### 2022-12-13 #5009 @editorialbot
[PRE REVIEW]: FitSNAP: Atomistic machine learning with LAMMPS
**Submitting author:** @rohskopf (<a href="http://orcid.org/0000-0002-2712-8296">Andrew Rohskopf</a>)
**Repository:** https://github.com/FitSNAP/FitSNAP
**Branch with paper.md** (empty if default branch):
**Version:** 3
