### 2022-12-13 9548ed6
made the path determining able to handle being passed a blank string as a way to indicate the local directory
