### 2022-12-13 828636f
added rough, unscaled DFT cost estimates to data saved for future plotting
