### 2022-12-13 fe2213d
fixed structures_chosen.dat not being written to the designated output directory
