### 2022-12-20 f02b0fd
bugfix to move creation of output directory to the head process only to avoid race condition
