### 2022-12-21 d131126
Write LAMMPS-ready pair network .pt file
