+++ fitsnap3lib/lib/neural_networks/write.py
@@ -1,5 +1,6 @@
 import numpy as np
 import torch
+import math
 
 def calc_n_params(model):
     """
@@ -159,7 +160,7 @@ class IgnoreElems(torch.nn.Module):
         super().__init__()
         self.subnet = subnet
 
-    def forward(self, descriptors, elems):
+    def forward(self, descriptors, elems=None):
         """
         Feeds descriptors to network model
 
@@ -311,3 +312,275 @@ class ElemwiseModels(torch.nn.Module):
             self.subnets[elem].to(self.dtype) 
             per_atom_attributes[elem_indices == i] = self.subnets[elem](descriptors[elem_indices == i]).flatten()
         return per_atom_attributes
+
+class PairNN(torch.nn.Module):
+    """
+    A class to wrap Modules to ensure lammps mliap compatability.
+    """
+
+    def __init__(self, model, n_descriptors, n_elements, n_params=None, device=None, dtype=torch.float64):
+        """
+        Constructs all the necessary attributes for the network module.
+        """
+
+        super().__init__()
+
+        print("^^^^^ write.py PairNN init")
+
+        self.model = model
+        self.device = device
+        self.dtype = dtype
+
+        # Put model on device and convert to dtype
+        self.to(self.dtype)
+        self.to(self.device)
+
+        if n_params is None:
+            n_params = calc_n_params(model)
+
+        self.n_params = n_params
+        self.n_descriptors = n_descriptors
+        self.n_elements = n_elements
+
+        # Parameters of model that will be used by LAMMPS:
+
+        self.cutoff = 5.0
+        self.num_radial_descriptors = 5
+        self.num_3body_descriptors = 12
+
+        # Parameters used by g3b:
+
+        self.pi = torch.tensor(math.pi)
+        self.eta = 4.
+        self.mu = torch.linspace(-1,1,self.num_3body_descriptors)
+
+        # trying to make a bessel object doesn't work:
+        #from fitsnap3lib.lib.neural_networks.descriptors.bessel import Bessel
+        #self.bessel = Bessel(self.num_radial_descriptors, self.cutoff) # Bessel object provides functions to calculate descriptors
+
+
+    def cutoff_function(self, r, numpy_bool=False):
+        """
+        Calculate cutoff function for all rij.
+
+        Args:
+            r (torch.Tensor.float): Array of pairwise distances with size (npairs, 1)
+
+        Returns:
+            function (torch.Tensor.float): Cutoff function values for all rij. 
+        """
+
+        rmin = 3.5
+
+        mask = r > rmin
+        if (r.dtype==torch.float64):
+            function = torch.empty(r.size()).double() # need double if doing FD test
+        else:
+            function = torch.empty(r.size())
+
+        c = self.cutoff
+        pi = torch.tensor(math.pi)
+
+        function[mask] = 0.5 + 0.5*torch.cos(pi*(r[mask]-rmin)/(c-rmin))
+        function[~mask] = 1.0
+
+        return function
+
+    def cutoff_function_g3b(self, rij):
+        """
+        Calculate cutoff function for all rij, used for g3b descriptors.
+
+        Args:
+            rij (:obj:`torch.Tensor`): Pairwise distances of all pairs in this batch, size 
+                                        (num_neighs, 1) where `num_neighs` is number of neighbors 
+                                        for all atoms in this batch.
+        """
+
+        c = self.cutoff
+        function = 0.5 + 0.5*torch.cos(self.pi*(rij-0)/(c-0))
+
+        return function[:,0]
+
+    def calculate_bessel(self, rij, cutoff_functions, n):
+        """
+        Calculate a specific radial bessel function `n` for all pairs.
+
+        Args:
+            rij (torch.Tensor.float): Pairwise distance tensor with size (num_neigh, 1).
[…]
