### 2022-12-22 a012d32
Working neighlist and energy agreement with LAMMPS
