### 2022-12-22 7c2da60
Working neighlist and energy agreement with LAMMPS
