### 2022-12-22 66e12ff
Correct pair NN force in LAMMPS
