### 2022-12-24 4d7c801
Working write.py force implementation
