### 2023-01-04 120a68f
Add optional shuffle flag for NN training
