### 2023-01-04 f2fd113
changed from recalculating entire DFT estimated cost each loop to just the amount added each time adding new structures
