### 2023-01-06 3518583
added pre-loop clustering and ability to grab the n_per_loop structures from separate clusters. also added an AL.in example file to the folder
