+++ examples/library/bayesian_active_learning/bayesian_active_learning.py
@@ -40,7 +40,10 @@ class AL_settings_class():
         self.active_learning = AL_configparser.getboolean('GENERAL', 'active_learning', fallback=True)
         self.number_of_iterations = AL_configparser.getint('GENERAL', 'number_of_iterations', fallback = 10)
         self.cluster_structures = AL_configparser.getboolean('GENERAL', 'cluster_structures', fallback = False)
-        self.batch_size = AL_configparser.getint('GENERAL', 'batch_size', fallback = 1)
+        #self.cluster_every_step = AL_configparser.getboolean('GENERAL', 'cluster_every_step', fallback = False)
+        # if wanting to implement, something like: pt.shared_arrays['a_copy'].array[mask_of_still_unused & (unlabeled_df['Row_Type']=='Energy')] as array to cluster
+        self.number_of_clusters = AL_configparser.getint('GENERAL', 'number_of_clusters', fallback = 0) #0 means to auto-determine number of clusters
+        self.batch_size = AL_configparser.getint('GENERAL', 'batch_size', fallback = 1) #0 can be used with clustering to take 1 structure from each cluster
         self.training_path = AL_configparser.get('GENERAL', 'training_path', fallback = None)
         self.unlabeled_path = AL_configparser.get('GENERAL', 'unlabeled_path', fallback = None)
         self.output_directory = AL_configparser.get('GENERAL', 'output_directory', fallback = getcwd())
@@ -247,7 +250,7 @@ def objective_function(df, EFS_reweighting=[1.0, 1.0, 1.0], FS_agg_functions=[No
     """
 
     if objective_function=='max': #simplest case - return the highest uncertainty row in each structure, taking the top nadd structures
-        ranked_structures = m_df.sort_values("uncertainty", ascending=False, key=abs).groupby(['Groups', 'Configs'], observed=True, sort=False).first()
+        ranked_structures = m_df.sort_values("uncertainty", ascending=False, key=abs).groupby(['Groups', 'Configs'], observed=True, sort=False).first() #first() could techinically get screwy if NaNs appear
         #x_vector_for_each_structure = ranked_structures[A_matrix_columns].values  ## TODO: will need to update how I'm grabbing this when aggregating F or S rows
     elif objective_function=='average':
         ranked_structures = m_df.groupby(['Groups', 'Configs'], observed=True, sort=False).agg({'uncertainty':np.average}).sort_values("uncertainty", ascending=False, key=abs)
@@ -263,6 +266,90 @@ def objective_function(df, EFS_reweighting=[1.0, 1.0, 1.0], FS_agg_functions=[No
     return ranked_structures#, x_vector_for_each_structure
 
 
+
+# K-means clustering, optionally using the elbow method for establishing best no. of clusters
+def eval_clustering(features, nc=None, need_centers=False, need_distances=False):
+    """
+        Evaluates clusters on a set of datapoints
+
+        inputs:
+        x_uq: some representation of datapoints - will be a matrix 'Energy' row of each structure or a transformation of it in this script
+        nc: number of clusters to make, if none, will determine number
+        need_distances: boolean (False), whether to return a matrix with the distance of each datapoint from each cluster center
+    
+        outputs:
+        icl: icl[i][:] is a torch tensor of local (x_uq) indices of data points in x_uq belonging to cluster i
+        distances (if need_distances=True): distance of each data point to each cluster center
+    """
+    scaler   = StandardScaler()
+    scaled_features = scaler.fit_transform(features)
+
+    # define parameters for the KMeans class setup
+    # n_init: the initial number of random cluster centers used in each iteration
+    # max_iter: number of iterations
[…]
