+++ examples/library/bayesian_active_learning/AL.in
@@ -1,0 +1,14 @@
+[GENERAL]
+active_learning =  True
+number_of_iterations= 10
+batch_size = 1
+
+[OBJECTIVE]
+objective_function = sum
+weight_by_relative_DFT_cost = True
+
+[VASP]
+
+[PLOTTING]
+plot_uncertainty_error_correlation = False
+plot_convergence_plots = True
