### 2023-01-06 3380baa
Debug PairNN writing class
