### 2023-01-06 f159ec4
Clean up pairnn
