+++ tools/xml2xyz.py
@@ -1,0 +1,243 @@
+"""
+This script parses through a single vasprun.xml file from VASP.
+This file may include one or more configurations.
+The output includes an XYZ file that can be used with FitSNAP.
+
+Usage:
+    python xml2xyz.py path/to/vasprun.xml path/to/file.xyz sample_every
+
+where sample_every specifies to collect data every that many timesteps.
+
+Alternatively, if you want to parse many vasprun.xml files in a directory, where each vasprun.xml
+contains a single configuration, do
+
+    python xml2xyz/py path/to/directory path/to/file.xyz sample_every
+
+where sample_every is now a meaningless parameter.
+Here, all data in the vasprun.xml files will be concatenated in a single file.xyz
+"""
+
+import sys, os
+import json
+from numpy import dot
+import xml.etree.ElementTree as ET
+
+xml_file = sys.argv[1]
+JSON_file = str(sys.argv[2])
+xyz_filename = str(sys.argv[2])
+sample_every = int(sys.argv[3]) # sample configurations every this many timesteps
+
+fh = open(xyz_filename, 'w')
+
+write_unconverged_steps_anyway = False
+
+def write_xyz(atom_coords, atom_force, stress_component, all_lattice, totalEnergy, 
+              listAtomTypes, natoms):
+    """
+    Write a single configuration to the XYZ file for this vasprun.xml file
+    """
+
+    print(f"natoms: {natoms}")
+
+    fh.write(f"{natoms}\n")
+    line = f'Lattice = "{all_lattice[0][0]} {all_lattice[0][1]} {all_lattice[0][2]} \
+                        {all_lattice[1][0]} {all_lattice[1][1]} {all_lattice[1][2]} \
+                        {all_lattice[2][0]} {all_lattice[2][1]} {all_lattice[2][2]}"'
+    line += " Properties=species:S:1:pos:R:3:forces:R:3 "
+    line += f"energy={totalEnergy} "
+    if (stress_component == []):
+        line += f'stress = "0 0 0 0 0 0 0 0 0"\n'
+    else:
+        line += f'stress = "{stress_component[0][0]} {stress_component[0][1]} {stress_component[0][2]} \
+                            {stress_component[1][0]} {stress_component[1][1]} {stress_component[1][2]} \
+                            {stress_component[2][0]} {stress_component[2][1]} {stress_component[2][2]}"\n'
+
+    fh.write(line)
+    
+    for n in range(0,natoms):
+        fh.write(f"{listAtomTypes[n]} {atom_coords[n][0]} {atom_coords[n][1]} {atom_coords[n][2]} {atom_force[n][0]} {atom_force[n][1]} {atom_force[n][2]}\n")
+
+
+def write_json(data, jsonfilename):
+    """
+    Define a function to take all the data that will be needed and send that to a
+    json.dump command which then takes that data and puts it into json format. 
+    The input for this function (data) is a dictionary that holds all the values from
+    the OUTCAR file that will be parsed through (such as atom positions, forces, lattice
+    vectors, etc.) which will be put into all Data with key 'Data'.  allDataHeader holds
+    all of the units as well as 'Data'.  This is then encompassed by myDataSet which then gets
+    fed into the json.dump command.  Most of this is just getting everything into the right format
+    
+    :param data: dict, parsed output of POSCAR and vasprun.xml files
+    :param jsonfilename: str, filename of .json file to be written 
+    """
+    jsonfile = open(jsonfilename, "w")
+    # print >>jsonfile, "# Header\n"
+    # print("# Header", file=jsonfilename)
+    allData = [data]
[…]
