### 2023-01-13 2c49b92
Use native Python int datatype since numpy int deprecated
