### 2023-01-16 e7f6816
changed when appending DFT costs to same time as appending error logs so that one list can not be longer than the others.
