### 2023-01-17 08a97f8
Clean up lammps model writer
