+++ fitsnap3lib/lib/neural_networks/write.py
@@ -518,23 +518,16 @@ class PairNN(torch.nn.Module):
         None
         """
 
-        #print("^^^^^ write.py pairnn forward")
-
-        #print("^^^^^ write.py rij:")
-        #print(rij)
-        #print(self.cutoff)
-
         rij = torch.from_numpy(rij).to(dtype=self.dtype, device=self.device).requires_grad_(True)
         unique_i = torch.from_numpy(unique_i).to(dtype=torch.long, device=self.device) #.requires_grad_(True)
         unique_j = torch.from_numpy(unique_j).to(dtype=torch.long, device=self.device) #.requires_grad_(True)
         tag_i = torch.from_numpy(tag_i).to(dtype=torch.long, device=self.device) #.requires_grad_(True)
         tag_j = torch.from_numpy(tag_j).to(dtype=torch.long, device=self.device) #.requires_grad_(True)
 
-        print(unique_i)
-        print(tag_i)
-        print(unique_j)
-        print(tag_j)
-        #assert(False)
+        #print(unique_i)
+        #print(tag_i)
+        #print(unique_j)
+        #print(tag_j)
 
         diff_norm = torch.nn.functional.normalize(rij, dim=1) # need for g3b
                                                               # size (npairs, 3)
@@ -543,10 +536,8 @@ class PairNN(torch.nn.Module):
 
         #print(distance_ij)
 
-        #print(unique_i)
-        #neighlist = torch.cat((unique_i, unique_j), dim=1)
-
-        maxr = torch.max(distance_ij)
+        # max pairwise distance is good for debugging match with fitting code
+        #maxr = torch.max(distance_ij)
         #print(f"Max pairwise distance: {maxr}")
 
         # Calculate cutoff functions once for pairwise terms here, because we use the same cutoff 
@@ -560,17 +551,19 @@ class PairNN(torch.nn.Module):
         rbf = self.radial_bessel_basis(distance_ij, cutoff_functions)
         assert(rbf.size()[0] == rij.size()[0])
 
+        # max rbf good for debugging match with fitting code
         #print(f"Max RBF: {torch.max(rbf)}")
 
         # calculate 3 body descriptors 
 
         descriptors_3body = self.calculate_g3b(distance_ij, diff_norm, unique_i)
 
+        # max d3b good for debugging match with fitting code
         #print(f"Max d3body: {torch.max(descriptors_3body)}")
 
         # concatenate radial descriptors and 3body descriptors
 
-        descriptors = torch.cat([rbf, descriptors_3body], dim=1) # num_pairs x num_descriptors
+        descriptors = torch.cat([rbf, descriptors_3body], dim=1) # size (num_pairs, num_descriptors)
         assert(descriptors.size()[0] == rij.size()[0])
 
         # input descriptors to a network for each pair; calculate pairwise energies
@@ -595,78 +588,3 @@ class PairNN(torch.nn.Module):
         #print(energy/54.)
         gradients_wrt_rij = torch.autograd.grad(energy, rij)[0]
         beta[:,:] = gradients_wrt_rij.detach().cpu().numpy().astype(np.float64)
-
-        # differentiate pairwise energies wrt rij
-        # NOTE: this doesn't work, gives (npair,3) derivative tensor but should be more like a Jacobian
-        #depair_drij = torch.autograd.grad(eij, 
-        #                        rij, 
-        #                        grad_outputs=torch.ones_like(eij))[0]
-
-        #print("Size of gradients_wrt_rij:")
-        #print(gradients_wrt_rij.size())
-
-        #print("Size of depair_drij:")
-        #print(depair_drij.size())
-
-        # calculate force on atom 1
[…]
