### 2023-01-17 9f063d7
Catch faulty xml files in xml2xyz converter
