+++ tools/xml2xyz.py
@@ -37,7 +37,7 @@ def write_xyz(atom_coords, atom_force, stress_component, all_lattice, totalEnerg
     Write a single configuration to the XYZ file for this vasprun.xml file
     """
 
-    print(f"natoms: {natoms}")
+    #print(f"natoms: {natoms}")
 
     fh.write(f"{natoms}\n")
     line = f'Lattice = "{all_lattice[0][0]} {all_lattice[0][1]} {all_lattice[0][2]} \
@@ -182,62 +182,68 @@ elif is_dir:
     for single_file in files:
         print(single_file)
         tree = ET.iterparse(single_file, events=['start', 'end'])
-        for event, elem in tree:
-            if elem.tag == 'parameters' and event=='end': #once at the start
-                NELM = int(elem.find('separator[@name="electronic"]/separator[@name="electronic convergence"]/i[@name="NELM"]').text)
-                #print(NELM)
-                
-            elif elem.tag == 'atominfo' and event == 'end': #once at the start
-                listAtomTypes = []
-                for entry in elem.find("array[@name='atoms']/set"):
-                    listAtomTypes.append(entry[0].text.strip())
-                natoms = len(listAtomTypes)
-                #print('atom types', listAtomTypes)
-                for entry in elem.find("array[@name='atomtypes']/set"):
-                    list_POTCARS.append(entry[4].text.strip().split())
-                #print('potcars:', list_POTCARS)
-                
-            elif (elem.tag == 'structure' and not elem.attrib.get('name')) and event=='end': #only the empty name ones - not primitive cell, initial, or final (those are repeats) - so each ionic step
-                all_lattice = []
-                for entry in elem.find("crystal/varray[@name='basis']"):
-                    lattice_row = [float(x) for x in entry.text.split()]
-                    all_lattice.append(lattice_row)
-                #print('lattice = ', all_lattice)
-                frac_atom_coords = []
-                for entry in elem.find("varray[@name='positions']"):
-                    frac_atom_coords.append([float(x) for x in entry.text.split()])
-                atom_coords = dot(frac_atom_coords, all_lattice).tolist()
-                print(atom_coords)
-                
-            elif elem.tag == 'calculation' and event=='end': #this triggers each ionic step
-                atom_force = []
-                force_block = elem.find("varray[@name='forces']")
-                if force_block:
-                    for entry in force_block:
-                        atom_force.append([float(x) for x in entry.text.split()])
-                print(atom_force)
-                stress_block = elem.find("varray[@name='stress']")
-                stress_component = []
-                if stress_block:
-                    for entry in stress_block:
-                        stress_component.append([float(x) for x in entry.text.split()])
-                totalEnergy = float(elem.find('energy/i[@name="e_0_energy"]').text)  ##NOTE! this value is incorrectly reported by VASP in version 5.4 (fixed in 6.1), see https://www.vasp.at/forum/viewtopic.php?t=17839
-                ## ASE vasprun.xml io reader has a more complex workaround to get the correct energy - we can update to include if needed
-                #print(totalEnergy)
-                if len(elem.findall("scstep")) == NELM:
-                    electronic_convergence = False ##This isn't the best way to check this, but not sure if info is directly available. Could try to calculate energy diff from scstep entries and compare to EDIFF
-                else:
-                    electronic_convergence = True
-                
-                #print(atom_coords)
+        try:
+            test = list(tree) # this will raise exception if faulty file
+            for event, elem in tree:
+                if elem.tag == 'parameters' and event=='end': #once at the start
+                    NELM = int(elem.find('separator[@name="electronic"]/separator[@name="electronic convergence"]/i[@name="NELM"]').text)
+                    #print(NELM)
+                    
+                elif elem.tag == 'atominfo' and event == 'end': #once at the start
+                    listAtomTypes = []
[…]
