### 2023-01-18 c264de4
Fix xml2xyz bug that caused file to not be written when reading faulty xml
