### 2023-01-18 #5009 @tpurcell90
Ya I can review this. Full disclosure I saw the project at a conference and recommended JOSS to one of the developers. I don't think that would be a conflict though.
