### 2023-01-18 #5009 @bjmorgan
I'm afraid I don't have any hands-on LAMMPS experience (although this submission looks like I might want to get some).
