### 2023-01-19 d8911ae
Don't shufle for ACE NN tests
