+++ tests/test_pytorch.py
@@ -249,6 +249,7 @@ def test_fd_ace_single_elem():
     #config.sections['ACE'].switchflag = 1 # required for smooth finite difference
     config.sections['PYTORCH'].manual_seed_flag = 1
     config.sections['PYTORCH'].dtype = torch.float64
+    config.sections['PYTORCH'].shuffle_flag = False # helps these finite difference tests
     # only perform calculations on displaced BCC structures
     config.sections['GROUPS'].group_table = {'Displaced_BCC': \
         {'training_size': 1.0, \
