+++ .github/workflows/pytests.yaml
@@ -59,6 +59,10 @@ jobs:
         # conda activate base
         cd ../
 
+        # increase postbuffer size
+        
+        git config --global http.postBuffer 157286400
+
         # Procedure for installing our custom branch with compute pace:
 
         git clone -b compute-pace https://github.com/jmgoff/lammps_compute_PACE
