### 2023-01-19 463f92f
Revert test git version to prevent error when cloning large lammps repo
