+++ .github/workflows/pytests.yaml
@@ -59,6 +59,12 @@ jobs:
         # conda activate base
         cd ../
 
+        # variables for git clone RPC error
+        
+        export GIT_TRACE_PACKET=1
+        export GIT_TRACE=1
+        export GIT_CURL_VERBOSE=1
+
         # increase postbuffer size
 
         git config --global http.version HTTP/1.1
