### 2023-01-19 4fe960b
Clean up pytorch model writer
