+++ fitsnap3lib/lib/neural_networks/write.py
@@ -6,11 +6,11 @@ def calc_n_params(model):
     """
     Returns the sum of two decimal numbers in binary digits.
 
-        Parameters:
-                model (torch.nn.Module): Network model that maps descriptors to a per atom attribute
+    Args:
+        model (torch.nn.Module): Network model that maps descriptors to a per atom attribute
 
-        Returns:
-                n_params (int): Number of NN model parameters
+    Returns:
+        n_params (int): Number of NN model parameters
     """
     return sum(p.nelement() for p in model.parameters())
 
@@ -18,59 +18,16 @@ class TorchWrapper(torch.nn.Module):
     """
     A class to wrap Modules to ensure lammps mliap compatability.
 
-    ...
-
-    Attributes
-    ----------
-    model : torch.nn.Module
-        Network model that maps descriptors to a per atom attribute
-
-    device : torch.nn.Module (None)
-        Accelerator device
-
-    dtype : torch.dtype (torch.float64)
-        Dtype to use on device
-
-    n_params : torch.nn.Module (None)
-        Number of NN model parameters
-
-    n_descriptors : int
-        Max number of per atom descriptors
-
-    n_elements : int
-        Max number of elements
-
-
-    Methods
-    -------
-    forward(descriptors, elems):
-        Feeds descriptors to network model to produce per atom energies and forces.
+    Attributes:
+        model (torch.nn.Module) : Network model that maps descriptors to a per atom attribute.
+        device (torch.nn.Module) : Accelerator device.
+        dtype (torch.dtype) : Data type to use on device. Default torch.float64.
+        n_params (int) : Number of NN model parameters.
+        n_descriptors (int) : Max number of per-atom descriptors.
+        n_elements (int) : Max number of element types.
     """
 
     def __init__(self, model, n_descriptors, n_elements, n_params=None, device=None, dtype=torch.float64):
-        """
-        Constructs all the necessary attributes for the network module.
-
-        Parameters
-        ----------
-            model : torch.nn.Module
-                Network model that maps descriptors to a per atom attribute
-
-            n_descriptors : int
-                Max number of per atom descriptors
-
-            n_elements : int
-                Max number of elements
-
-            n_params : torch.nn.Module (None)
-                Number of NN model parameters
-            
-            device : torch.nn.Module (None)
-                Accelerator device
-            
-            dtype : torch.dtype (torch.float64)
-                Dtype to use on device
-        """
 
         super().__init__()
 
@@ -94,31 +51,18 @@ class TorchWrapper(torch.nn.Module):
         Takes element types and descriptors calculated via lammps and
         calculates the per atom energies and forces.
 
-        Parameters
-        ----------
-        elems : numpy.array
-            Per atom element types, starting from -1
+        Args:
+            elems (np.array) : Per-atom element types. Used to start from -1.
+            descriptors (np.array) : Per-atom descriptors.
+            beta (np.array) : Expired beta array to be filled with new betas.
+            energy (np.array) : Expired per-atom energy array to be filled with new per-atom energies.
 
-        descriptors : numpy.array
-            Per atom descriptors
-
-        beta : numpy.array
-            Expired beta array to be filled with new betas
-
-        energy : numpy.array
-            Expired per atom energy array to be filled with new per atom energy
-            (Note: This is a pointer to the lammps per atom energies)
-
-
-        Returns
-        -------
-        None
+        Returns:
+            No return.
         """
 
         descriptors = torch.from_numpy(descriptors).to(dtype=self.dtype, device=self.device).requires_grad_(True)
         elems = torch.from_numpy(elems).to(dtype=torch.long, device=self.device)
-        #print(elems)
[…]
