+++ fitsnap3lib/lib/neural_networks/pytorch.py
@@ -142,10 +142,6 @@ class FitTorch(torch.nn.Module):
         else:
             predicted_energy_total = None
 
-
-        #print(predicted_energy_total/192)
-        #assert(False)
-
         # calculate forces
 
         if (self.force_bool):
