### 2023-01-19 73cdb34
putting a couple more flags into the example AL.in with their default setting
