+++ examples/library/bayesian_active_learning/AL.in
@@ -2,6 +2,9 @@
 active_learning =  True
 number_of_iterations= 10
 batch_size = 1
+cluster_structures = False
+number_of_clusters = 0
+
 
 [OBJECTIVE]
 objective_function = sum
