### 2023-01-19 171e159
set tracking of DFT estimated costs as optional and defaulting to off
