+++ examples/library/bayesian_active_learning/AL.in
@@ -16,3 +16,5 @@ weight_by_relative_DFT_cost = True
 plot_uncertainty_error_correlation = False
 plot_convergence_plots = True
 track_estimated_DFT_cost = False
+
+
