+++ tools/xml2xyz.py
@@ -15,15 +15,15 @@ vasprun.xml contains a single configuration, do
 
 where sample_every is now a meaningless parameter.
 Here, all data in the vasprun.xml files will be concatenated in a single file.xyz
+
+TODO: Merge this script into the xml2JSON converter.
 """
 
 import sys, os
-import json
 from numpy import dot
 import xml.etree.ElementTree as ET
 
 xml_file = sys.argv[1]
-JSON_file = str(sys.argv[2])
 xyz_filename = str(sys.argv[2])
 sample_every = int(sys.argv[3]) # sample configurations every this many timesteps
 
@@ -57,47 +57,10 @@ def write_xyz(atom_coords, atom_force, stress_component, all_lattice, totalEnerg
     for n in range(0,natoms):
         fh.write(f"{listAtomTypes[n]} {atom_coords[n][0]} {atom_coords[n][1]} {atom_coords[n][2]} {atom_force[n][0]} {atom_force[n][1]} {atom_force[n][2]}\n")
 
-
-def write_json(data, jsonfilename):
-    """
-    Define a function to take all the data that will be needed and send that to a
-    json.dump command which then takes that data and puts it into json format. 
-    The input for this function (data) is a dictionary that holds all the values from
-    the OUTCAR file that will be parsed through (such as atom positions, forces, lattice
-    vectors, etc.) which will be put into all Data with key 'Data'.  allDataHeader holds
-    all of the units as well as 'Data'.  This is then encompassed by myDataSet which then gets
-    fed into the json.dump command.  Most of this is just getting everything into the right format
-    
-    :param data: dict, parsed output of POSCAR and vasprun.xml files
-    :param jsonfilename: str, filename of .json file to be written 
-    """
-    jsonfile = open(jsonfilename, "w")
-    # print >>jsonfile, "# Header\n"
-    # print("# Header", file=jsonfilename)
-    allData = [data]
-    allDataHeader = {}
-    allDataHeader["EnergyStyle"] = "electronvolt"
-    allDataHeader["StressStyle"] = "kB"
-    allDataHeader["AtomTypeStyle"] = "chemicalsymbol"
-    allDataHeader["PositionsStyle"] = "angstrom"
-    allDataHeader["ForcesStyle"] = "electronvoltperangstrom"
-    allDataHeader["LatticeStyle"] = "angstrom"
-    allDataHeader["Data"] = allData
-
-    myDataset = {}
-
-    myDataset["Dataset"] = allDataHeader
-    #jsonfile.write(json.dumps(myDataset))  # if you want a condensed string
-    json.dump(myDataset, jsonfile, indent=2, sort_keys=True)  #if you want the expanded, multi-line format
-    jsonfile.close()
-    return
-
-
 order_atom_types = []
 listAtomTypes = []
 list_POTCARS = []
 config_number = 1
-
 
 # Start parsing through vasprun.xml looking for entries that are associated with the
 # different values for the data needed, such as forces or positions
@@ -116,27 +79,22 @@ if (is_file):
     for event, elem in tree:
         if elem.tag == 'parameters' and event=='end': #once at the start
             NELM = int(elem.find('separator[@name="electronic"]/separator[@name="electronic convergence"]/i[@name="NELM"]').text)
-            #print(NELM)
         elif elem.tag == 'atominfo' and event == 'end': #once at the start
             for entry in elem.find("array[@name='atoms']/set"):
                 listAtomTypes.append(entry[0].text.strip())
             natoms = len(listAtomTypes)
-            #print('atom types', listAtomTypes)
             for entry in elem.find("array[@name='atomtypes']/set"):
                 list_POTCARS.append(entry[4].text.strip().split())
-            #print('potcars:', list_POTCARS)
             
[…]
