### 2023-01-20 cde4416
Write less spaces for lattice string in xyz files
