### 2023-01-20 6180d20
minor edits to inline documentation for active learning script
