### 2023-01-20 3145d94
bugfix to allow plotting uncertainty vs error plots when using bzeroflag
