### 2023-01-20 3fc9370
bugfix for the bugfix because I should fully read numpy convenience functions before using them
