### 2023-01-21 8aafe17
added some commented out code used to plot group separated error-uncertainty plots in case it is useful in the future
