### 2023-01-22 #5009 @rohskopf
Thanks @jmschrei and @tpurcell90.
@bjmorgan no LAMMPS experience required; our tutorial in Google Colab installs everything and walks through examples.
