### 2023-01-26 35bbac8
updated documentation at top of AL script and allowed for plotting of all test data if no testing_json_group - not yet made plotting for no testing at all
