### 2023-01-26 22b88b5
added support for writing output every n loops and made fitsnap models be copied to the output directory
