### 2023-01-27 7c605e9
bugfixing my sloppily written copy commands
