### 2023-01-27 208a676
changed copy2 to move commands to avoid extra files lying around and added a check to prevent trying to move a file to its own location when output directory is the current working directory
