### 2023-02-02 37061e1
Use proper hybrid snap style and add atomic style
