### 2023-02-02 5bab2c5
ID fits and associated files with a hash
