+++ fitsnap3lib/io/outputs/snap.py
@@ -3,6 +3,7 @@ from fitsnap3lib.parallel_tools import ParallelTools
 from datetime import datetime
 from fitsnap3lib.io.input import Config
 import numpy as np
+import random
 
 
 #config = Config()
@@ -85,20 +86,28 @@ def _to_param_string():
         chemflag_int = 1
     else:
         chemflag_int = 0
-    return f"""
-    # required
-    rcutfac {config.sections["BISPECTRUM"].rcutfac}
-    twojmax {max(config.sections["BISPECTRUM"].twojmax)}
+    # Report reference section settings
+    unit = f"# units {config.sections['REFERENCE'].units}\n"
+    atom = f"# atom_style {config.sections['REFERENCE'].atom_style}\n"
+    pair = "\n".join(["# " + s for s in config.sections["REFERENCE"].lmp_pairdecl]) + "\n"
+    refsec = unit + atom + pair
+    # Build output for param file
+    out = "# required\n"
+    out += f"rcutfac {config.sections['BISPECTRUM'].rcutfac}\n"
+    out += f"twojmax {max(config.sections['BISPECTRUM'].twojmax)}\n\n"
+    out += "# optional\n"
+    out += f"rcutfac {config.sections['BISPECTRUM'].rfac0}\n"
+    out += f"rmin0 {config.sections['BISPECTRUM'].rmin0}\n"
+    out += f"bzeroflag {config.sections['BISPECTRUM'].bzeroflag}\n"
+    out += f"wselfallflag {config.sections['BISPECTRUM'].wselfallflag}\n"
+    out += f"chemflag {chemflag_int}\n"
+    out += f"bnormflag {config.sections['BISPECTRUM'].bnormflag}\n\n"
+    out += "# This file was generated by FitSNAP.\n"
+    out += f"# Hash: {config.hash}\n"
+    out += "# FitSNAP REFERENCE section settings:\n"
+    out += f"{refsec}"
 
-    #  optional
-    rfac0 {config.sections["BISPECTRUM"].rfac0}
-    rmin0 {config.sections["BISPECTRUM"].rmin0}
-    bzeroflag {config.sections["BISPECTRUM"].bzeroflag}
-    quadraticflag {config.sections["BISPECTRUM"].quadraticflag}
-    wselfallflag {config.sections["BISPECTRUM"].wselfallflag}
-    chemflag {chemflag_int}
-    bnormflag {config.sections["BISPECTRUM"].bnormflag}
-    """
+    return out
 
 
 def _to_coeff_string(coeffs):
@@ -170,7 +179,7 @@ def _to_lammps_input():
     """
     Use config settings to write a LAMMPS input script.
     """
-    
+
     config = Config()
 
     snap_filename = config.sections["OUTFILE"].potential_name.split("/")[-1]
