### 2023-02-03 cecf206
Package fit files into tarball
