+++ fitsnap3lib/io/outputs/snap.py
@@ -4,6 +4,7 @@ from datetime import datetime
 from fitsnap3lib.io.input import Config
 import numpy as np
 import random
+import tarfile
 
 
 #config = Config()
@@ -42,7 +43,14 @@ class Snap(Output):
                 file.write(_to_potential_file())
             with optional_open("in.lammps", 'wt') as file:
                 file.write(_to_lammps_input())
-
+            # Add these files to a tarball
+            fp = tarfile.open(f"fit-{self.config.hash}.tar.gz", 'w:gz')
+            potname = self.config.sections["OUTFILE"].potential_name.split('/')[-1]
+            fp.add(potname + '.snapcoeff')
+            fp.add(potname + '.snapparam')
+            fp.add(potname + '.mod')
+            fp.add("in.lammps")
+            fp.close()
             self.write_errors(errors)
         decorated_write()
 
@@ -109,7 +117,6 @@ def _to_param_string():
 
     return out
 
-
 def _to_coeff_string(coeffs):
     """
     Convert a set of coefficients along with bispec options into a .snapparam file
@@ -124,7 +131,7 @@ def _to_coeff_string(coeffs):
         blank2Js = np.insert(blank2Js, 0, [1.0], axis=1)
     coeffs = np.multiply(coeffs, blank2Js)
     type_names = config.sections["BISPECTRUM"].types
-    out = f"# fitsnap fit generated on {datetime.now()}\n\n"
+    out = f"# FitSNAP fit generated on {datetime.now()} with Hash: {config.hash}\n\n"
     out += "{} {}\n".format(len(type_names), ncoeff+1)
 
     for elname, rjval, wjval, column, ielem in zip(type_names,
@@ -143,12 +150,15 @@ def _to_coeff_string(coeffs):
 
 def _to_potential_file():
     """
-    Use config settings to write a LAMMPS potential file.
+    Use config settings to write a LAMMPS potential .mod file.
     """
     config = Config()
 
     ps = config.sections["REFERENCE"].lmp_pairdecl[0]
     snap_filename = config.sections["OUTFILE"].potential_name.split("/")[-1]
+
+    out = "# This file was generated by FitSNAP.\n"
+    out += f"# Hash: {config.hash}\n\n"
 
     if "hybrid" in ps:
         # extract non zero parts of pair style
@@ -158,7 +168,7 @@ def _to_potential_file():
             del split[zero_indx]
             del split[zero_indx] # delete the zero pair cutoff
             ps = ' '.join(split)
-        out = ps + " snap\n"
+        out += ps + " snap\n"
         # add pair coeff commands from input, ignore if pair zero
         for pc in config.sections["REFERENCE"].lmp_pairdecl[1:]:
             out += f"{pc}\n" if "zero" not in pc else ""
@@ -167,7 +177,7 @@ def _to_potential_file():
             pc_snap += f" {t}"
         out += pc_snap
     else:
-        out = "pair_style snap\n"
+        out += "pair_style snap\n"
         pc_snap = f"pair_coeff * * {snap_filename}.snapcoeff {snap_filename}.snapparam"
         for t in config.sections["BISPECTRUM"].types:
             pc_snap += f" {t}"
