### 2023-02-04 215c740
Reformat Running FitSNAP docs and document cmdline options for #169
