+++ docs/source/Run/Run_outputs.rst
@@ -1,0 +1,12 @@
+FitSNAP Outputs
+===============
+
+FitSNAP outputs include error metrics, detailed errors for each atom, configuration, or groups of 
+configurations, and LAMMPS-ready files for running MD simulations.
+
+Outputs are different for linear and nonlinear models. 
+
+For linear models, please see the `Linear models output section <Linear.html#outputs>`__
+
+For nonlinear models, please see the
+`PyTorch models output section <Pytorch.html#outputs-and-error-calculation>`__.
