### 2023-02-04 cf80fa0
Write switchinnerflag settings
Co-authored-by: megmcca <megmcca@sandia.gov>
