+++ fitsnap3lib/io/outputs/snap.py
@@ -112,7 +112,14 @@ def _to_param_string():
     out += f"bzeroflag {config.sections['BISPECTRUM'].bzeroflag}\n"
     out += f"wselfallflag {config.sections['BISPECTRUM'].wselfallflag}\n"
     out += f"chemflag {chemflag_int}\n"
-    out += f"bnormflag {config.sections['BISPECTRUM'].bnormflag}\n\n"
+    out += f"bnormflag {config.sections['BISPECTRUM'].bnormflag}\n"
+    out += f"switchinnerflag {config.sections['BISPECTRUM'].switchinnerflag}\n"
+    if config.sections["BISPECTRUM"].switchinnerflag:
+        sinner = config.sections["BISPECTRUM"].sinner
+        dinner = config.sections["BISPECTRUM"].dinner
+        out += f"sinner {sinner}\n"
+        out += f"dinner {dinner}\n"
+    out += "\n"
     out += "# This file was generated by FitSNAP.\n"
     out += f"# Hash: {config.hash}\n"
     out += "# FitSNAP REFERENCE section settings:\n"
