### 2023-02-04 94bbd79
Write NN-SNAP LAMMPS files
