+++ fitsnap3lib/io/outputs/snap.py
@@ -19,11 +19,16 @@ class Snap(Output):
         self.pt = ParallelTools()
 
     def output(self, coeffs, errors):
-        new_coeffs = None
-        # new_coeffs = pt.combine_coeffs(coeffs)
-        if new_coeffs is not None:
-            coeffs = new_coeffs
-        self.write(coeffs, errors)
+        if (self.config.sections["CALCULATOR"].nonlinear):
+            # Currently coeffs and errors are empty for nonlinear.
+            # TODO: Add nonlinear error calculation for output here, similar format as linear.
+            self.write_nn(errors)
+        else:
+            new_coeffs = None
+            # new_coeffs = pt.combine_coeffs(coeffs)
+            if new_coeffs is not None:
+                coeffs = new_coeffs
+            self.write(coeffs, errors)
 
     #@pt.rank_zero
     def write(self, coeffs, errors):
@@ -55,6 +60,23 @@ class Snap(Output):
                 fp.close()
 
             self.write_errors(errors)
+        decorated_write()
+
+    def write_nn(self, errors):
+        """ Write output for nonlinear fits. """
+        # TODO: Currently errors is empty; need to add error calculation and write same format 
+        #       file as linear.
+        @self.pt.rank_zero
+        def decorated_write():
+            if self.config.sections["EXTRAS"].only_test != 1:
+                if self.config.sections["CALCULATOR"].calculator != "LAMMPSSNAP":
+                    raise TypeError("SNAP output style must be paired with LAMMPSSNAP calculator")
+            with optional_open(self.config.sections["OUTFILE"].potential_name and
+                               self.config.sections["OUTFILE"].potential_name + '.mliap.descriptor', 'wt') as file:
+                file.write(_to_mliap_string())
+            with optional_open(self.config.sections["OUTFILE"].potential_name and
+                               self.config.sections["OUTFILE"].potential_name + '.mod', 'wt') as file:
+                file.write(_to_mliap_mod())
         decorated_write()
 
     #@pt.sub_rank_zero
@@ -115,10 +137,8 @@ def _to_param_string():
     out += f"bnormflag {config.sections['BISPECTRUM'].bnormflag}\n"
     out += f"switchinnerflag {config.sections['BISPECTRUM'].switchinnerflag}\n"
     if config.sections["BISPECTRUM"].switchinnerflag:
-        sinner = config.sections["BISPECTRUM"].sinner
-        dinner = config.sections["BISPECTRUM"].dinner
-        out += f"sinner {sinner}\n"
-        out += f"dinner {dinner}\n"
+        out += f"sinner {config.sections['BISPECTRUM'].sinner}\n"
+        out += f"dinner {config.sections['BISPECTRUM'].dinner}\n"
     out += "\n"
     out += "# This file was generated by FitSNAP.\n"
     out += f"# Hash: {config.hash}\n"
@@ -234,5 +254,73 @@ def _to_lammps_input():
     out += "# Run dynamics\n"
     out += "\n"
     out += "run 1000\n"
+
+    return out
+
+def _to_mliap_string():
+    """ Build mliap descriptor file. """
+    config = Config()
+    out = "# required\n"
+    out += f"rcutfac {config.sections['BISPECTRUM'].rcutfac}\n"
+    out += f"twojmax {max(config.sections['BISPECTRUM'].twojmax)}\n\n"
+    out += "#elements\n"
+    out += f"nelems {config.sections['BISPECTRUM'].numtypes}\n"
+    out += "elems"
+    for t in config.sections['BISPECTRUM'].types:
[…]
