### 2023-02-04 b8a950d
Use f-string in mliap pair style writing
