### 2023-02-05 bc4fe62
Add LAMMPS Python library examples of evaluating energies/forces with trained model
