### 2023-02-06 80fa040
Uncomment relative cmd line option since it is working
