### 2023-02-06 43b6798
Clean up old commented output for nonlinear models
