+++ fitsnap3lib/fitsnap.py
@@ -157,11 +157,5 @@ class FitSnap:
         def decorated_write_output():
             if not self.config.args.perform_fit:
                 return
-            # prevent Output class from trying to process non-existing solver.fit when using nonlinear solvers
             output.output(self.solver.fit, self.solver.errors)
-            #if self.solver.linear:
-            #    output.output(self.solver.fit, self.solver.errors)
-            #else:
-            #    if (self.pt._rank==0):
-            #        print("WARNING: Nonlinear solvers do not output generic output - see PyTorch/JAX model files.")
         decorated_write_output()
