+++ fitsnap3lib/io/outputs/snap.py
@@ -276,6 +276,10 @@ def _to_mliap_string():
     out += "welems"
     for t in range(config.sections['BISPECTRUM'].numtypes):
         out += f" {config.sections['BISPECTRUM'].wj[t]}"
+    out += "\n"
+    if config.sections["BISPECTRUM"].switchinnerflag:
+        out += f"sinnerelems {config.sections['BISPECTRUM'].sinner}\n"
+        out += f"dinnerelems {config.sections['BISPECTRUM'].dinner}\n"
     out += "\n\n"
     out += "# optional\n"
     out += f"rcutfac {config.sections['BISPECTRUM'].rfac0}\n"
