### 2023-02-09 eb455a5
Use switchinnerflag in NN output
