### 2023-02-10 8eb1204
Begin NN error analysis
