+++ fitsnap3lib/solvers/solver.py
@@ -110,7 +110,38 @@ class Solver:
         @self.pt.rank_zero
         def decorated_error_analysis():
             if not self.linear:
-                self.pt.single_print("No Error Analysis for non-linear potentials")
+                self.pt.single_print("NN error analysis; using configuration class.")
+                (energies_model, forces_model) = self.evaluate_configs(option=1, standardize_bool=False)
+                print(len(energies_model))
+                print(len(forces_model))
+                #self.df = DataFrame(self.pt.shared_arrays['a'].array)
+                #print(self.df)
+                # make a list of per-atom quantities like type, i, forces, etc
+                #fx_pred = [self.configs[m].forces[3*i+0] for i in range(self.configs[m].natoms) for m in range(len(self.configs))]
+                fha = open("per-atom.dat", 'w')
+                fhc = open("per-config.dat", 'w')
+                atom_indx = 0
+                m = 0
+                for c in self.configs:
+                    e_pred = energies_model[m].detach().numpy()
+                    f_pred = forces_model[m].detach().numpy()
+                    print(e_pred)
+                    for i in range(c.natoms):
+                        fx_truth = c.forces[3*i+0]
+                        fy_truth = c.forces[3*i+1]
+                        fz_truth = c.forces[3*i+2]
+                        fx_pred = f_pred[3*i+0]
+                        fy_pred = f_pred[3*i+1]
+                        fz_pred = f_pred[3*i+2]
+                        line = f"{fx_truth} {fy_truth} {fz_truth} "
+                        line += f"{fx_pred} {fy_pred} {fz_pred}"
+                        fha.write(line + "\n")
+                        atom_indx += 1
+                    m += 1
+                fha.close()
+                fhc.close()
+                #print(self.configs)
+
                 return
 
             # collect remaining arrays to write dataframe
