### 2023-02-10 53717df
Write detailed NN fit info to per-atom and per-config files
