+++ fitsnap3lib/solvers/solver.py
@@ -76,7 +76,8 @@ class Solver:
 
     def _ncount_mae_rmse_rsq_unweighted_and_weighted(self, g):
         """
-        Calculate errors given a dataframe.
+        Calculate errors given a dataframe. The name of this function denotes the quantities it 
+        returns as a Pandas series.
 
         Args:
             g: Pandas dataframe
@@ -110,14 +111,76 @@ class Solver:
         @self.pt.rank_zero
         def decorated_error_analysis():
             if not self.linear:
+                mae_f = {} # Force MAE of each group, train and test.
+                mae_e = {} # Test energy MAE of each group, train and test.
+                rmse_f = {} # Test force RMSE of each group, train and test.
+                rmse_e = {} # Test energy RMSE of each group, train and test.
+
+                count_train = {} # Nested dictionary with number of configs and atoms per group for training data.
+                count_test = {} # Nested dictionary with number of configs and atoms per group for testing data.
+                for group in self.config.sections['GROUPS'].group_table:
+                    mae_f[group] = {}
+                    mae_e[group] = {}
+                    rmse_f[group] = {}
+                    rmse_e[group] = {}
+
+                    mae_f[group]["train"],  mae_f[group]["test"] =  0.0, 0.0
+                    mae_e[group]["train"],  mae_e[group]["test"] =  0.0, 0.0
+                    rmse_f[group]["train"], rmse_f[group]["test"] = 0.0, 0.0
+                    rmse_e[group]["train"], rmse_e[group]["test"] = 0.0, 0.0
+
+                    count_test[group] = {}
+                    count_test[group]["nconfigs"] = 0 # Total number test configs in group.
+                    count_test[group]["natoms"] = 0 # Total number test atoms in group.
+                    count_train[group] = {}
+                    count_train[group]["nconfigs"] = 0 # Total number test configs in group.
+                    count_train[group]["natoms"] = 0 # Total number test atoms in group.
+
+                # Add dictionary keys for total data '*ALL'
+
+                mae_f['*ALL'] = {}
+                mae_e['*ALL'] = {}
+                rmse_f['*ALL'] = {}
+                rmse_e['*ALL'] = {}
+                mae_f['*ALL']["train"],  mae_f['*ALL']["test"] =  0.0, 0.0
+                mae_e['*ALL']["train"],  mae_e['*ALL']["test"] =  0.0, 0.0
+                rmse_f['*ALL']["train"], rmse_f['*ALL']["test"] = 0.0, 0.0
+                rmse_e['*ALL']["train"], rmse_e['*ALL']["test"] = 0.0, 0.0
+                count_test['*ALL'] = {}
+                count_test['*ALL']["nconfigs"] = 0 # Total number test configs in group.
+                count_test['*ALL']["natoms"] = 0 # Total number test atoms in group.
+                count_train['*ALL'] = {}
+                count_train['*ALL']["nconfigs"] = 0 # Total number test configs in group.
+                count_train['*ALL']["natoms"] = 0 # Total number test atoms in group.
+
                 self.pt.single_print("NN error analysis; using configuration class.")
+
                 (energies_model, forces_model) = self.evaluate_configs(option=1, standardize_bool=False)
                 fha = open("peratom.dat", 'w')
                 fhc = open("perconfig.dat", 'w')
                 atom_indx = 0
                 m = 0
                 for c in self.configs:
-                    e_pred = energies_model[m].detach().numpy()
+                    e_pred = energies_model[m].detach().numpy()/c.natoms # Model per-atom energy.
+
+                    ae = abs(c.energy - e_pred)
[…]
