### 2023-02-13 95f2c06
Fix rcutfac
Co-authored-by: elsikorski <embersikorski@u.boisestate.edu>
