+++ fitsnap3lib/io/outputs/snap.py
@@ -129,7 +129,7 @@ def _to_param_string():
     out += f"rcutfac {config.sections['BISPECTRUM'].rcutfac}\n"
     out += f"twojmax {max(config.sections['BISPECTRUM'].twojmax)}\n\n"
     out += "# optional\n"
-    out += f"rcutfac {config.sections['BISPECTRUM'].rfac0}\n"
+    out += f"rfac0 {config.sections['BISPECTRUM'].rfac0}\n"
     out += f"rmin0 {config.sections['BISPECTRUM'].rmin0}\n"
     out += f"bzeroflag {config.sections['BISPECTRUM'].bzeroflag}\n"
     out += f"wselfallflag {config.sections['BISPECTRUM'].wselfallflag}\n"
