### 2023-02-14 6bbfc79
Optionalize peratom and perconfig data writing
