### 2023-02-14 55d3b05
Allow zero fraction training/testing in new NN output
