+++ fitsnap3lib/solvers/solver.py
@@ -234,35 +234,40 @@ class Solver:
                 #print(self.configs)
 
                 # Normalize to get average errors.
-                mae_f['*ALL']["test"] /= 3*count_test['*ALL']["natoms"]
-                mae_f['*ALL']["train"] /= 3*count_train['*ALL']["natoms"]
-                rmse_f['*ALL']["test"] /= 3*count_test['*ALL']["natoms"]
-                rmse_f['*ALL']["test"] /= np.sqrt(rmse_f['*ALL']["test"])
-                rmse_f['*ALL']["train"] /= 3*count_train['*ALL']["natoms"]
-                rmse_f['*ALL']["train"] /= np.sqrt(rmse_f['*ALL']["train"])
-                mae_e['*ALL']["test"] /= count_test['*ALL']["nconfigs"]
-                mae_e['*ALL']["train"] /= count_train['*ALL']["nconfigs"]
-                rmse_e['*ALL']["test"] /= count_test['*ALL']["nconfigs"]
-                rmse_e['*ALL']["test"] = np.sqrt(rmse_e['*ALL']["test"])
-                rmse_e['*ALL']["train"] /= count_train['*ALL']["nconfigs"]
-                rmse_e['*ALL']["train"] = np.sqrt(rmse_e['*ALL']["train"])
+
+                # Force MAE.
+                mae_f['*ALL']["test"]   /= 3*count_test['*ALL']["natoms"]  if count_test[group]["natoms"]   > 0 else np.nan
+                mae_f['*ALL']["train"]  /= 3*count_train['*ALL']["natoms"] if count_train[group]["natoms"]  > 0 else np.nan
+                # Force RMSE. 
+                rmse_f['*ALL']["test"]  /= 3*count_test['*ALL']["natoms"]  if count_test[group]["natoms"]   > 0 else np.nan
+                rmse_f['*ALL']["test"]   = np.sqrt(rmse_f['*ALL']["test"])
+                rmse_f['*ALL']["train"] /= 3*count_train['*ALL']["natoms"] if count_train[group]["natoms"]  > 0 else np.nan
+                rmse_f['*ALL']["train"]  = np.sqrt(rmse_f['*ALL']["train"])
+                # Energy MAE.
+                mae_e['*ALL']["test"]   /= count_test['*ALL']["nconfigs"]  if count_test[group]["nconfigs"]  > 0 else np.nan
+                mae_e['*ALL']["train"]  /= count_train['*ALL']["nconfigs"] if count_train[group]["nconfigs"] > 0 else np.nan
+                # Energy RMSE.
+                rmse_e['*ALL']["test"]  /= count_test['*ALL']["nconfigs"]  if count_test[group]["nconfigs"]  > 0 else np.nan
+                rmse_e['*ALL']["test"]   = np.sqrt(rmse_e['*ALL']["test"])
+                rmse_e['*ALL']["train"] /= count_train['*ALL']["nconfigs"] if count_train[group]["nconfigs"] > 0 else np.nan
+                rmse_e['*ALL']["train"]  = np.sqrt(rmse_e['*ALL']["train"])
                 for group in self.config.sections['GROUPS'].group_table:
                     # Force MAE
-                    mae_f[group]["test"] /= 3*count_test[group]["natoms"]
-                    mae_f[group]["train"] /= 3*count_train[group]["natoms"]
+                    mae_f[group]["test"]   /= 3*count_test[group]["natoms"]  if count_test[group]["natoms"]    > 0  else np.nan
+                    mae_f[group]["train"]  /= 3*count_train[group]["natoms"] if count_train[group]["natoms"]   > 0  else np.nan
                     # Force RMSE
-                    rmse_f[group]["test"] /= 3*count_test[group]["natoms"]
[…]
