### 2023-02-14 efc6711
Confirm NN error analysis with external script
