+++ fitsnap3lib/tools/nn_tools.py
@@ -1,0 +1,154 @@
+import numpy as np
+import pandas as pd
+from itertools import compress
+import matplotlib.pyplot as plt
+import matplotlib.patches as mpatches
+
+SMALL_SIZE = 10
+MEDIUM_SIZE = 12
+BIGGER_SIZE = 12
+plt.rc('font', size=SMALL_SIZE)          # controls default text sizes
+plt.rc('axes', titlesize=SMALL_SIZE)     # fontsize of the axes title
+plt.rc('axes', labelsize=MEDIUM_SIZE)    # fontsize of the x and y labels
+plt.rc('xtick', labelsize=SMALL_SIZE)    # fontsize of the tick labels
+plt.rc('ytick', labelsize=SMALL_SIZE)    # fontsize of the tick labels
+plt.rc('legend', fontsize=SMALL_SIZE)    # legend fontsize
+plt.rc('figure', titlesize=BIGGER_SIZE)  # fontsize of the figure title
+
+class NNTools():
+    """
+    Class containing methods that help process output for NN models.
+
+    Args:
+        peratom_file : Path to a per atom data file from FitSNAP NN output. Defaults to None.
+        perconfig_file : Path to a per config data file from FitSNAP NN output. Defaults to None.
+
+    Attributes:
+        dfa : Dataframe of per atom data.
+        dfc : Dataframe of per config data.
+
+    """
+    def __init__(self, peratom_file: str = None, perconfig_file: str = None):
+        self.dfa = pd.read_csv(peratom_file,
+                   delimiter = ' ',
+                   lineterminator = '\n',
+                   header = 0) if peratom_file else None
+        self.dfc = pd.read_csv(perconfig_file,
+                   delimiter = ' ',
+                   lineterminator = '\n',
+                   header = 0) if peratom_file else None
+
+    def calc_errors(self, quantity: str) -> dict:
+        """
+        Calculate errors using FitSNAP NN output files.
+
+        Args:
+            quantity : "Energy" or "Force".
+
+        Returns:
+            Tuple of error values (mae_train, mae_test, rmse_train, rmse_test) for given a quantity.
+            Dictionary of error values for given quantity. First key is error metric, second key is 
+            train/test. E.g. ["test"]["mae"] denotes the test MAE for a given quantity.
+        """
+
+        ret = {}
+        ret["train"] = {}
+        ret["test"] = {}
+        if (quantity == "Energy"):
+            test_bool = self.dfc.loc[:, "Testing_Bool"].tolist()
+            test_idx = [i for i, x in enumerate(test_bool) if x]
+            train_idx = [i for i, x in enumerate(test_bool) if not x]
+            # Extract train energies.
+            true_train = np.array(self.dfc.loc[train_idx, "Energy_Truth"].tolist())
+            pred_train = np.array(self.dfc.loc[train_idx, "Energy_Pred"].tolist())
+            ret["train"]["mae"] = np.mean(abs(true_train - pred_train))
+            ret["train"]["rmse"] = np.sqrt(np.mean(np.square(true_train - pred_train)))
+            # Extract test energies.
+            true_test = np.array(self.dfc.loc[test_idx, "Energy_Truth"].tolist())
+            pred_test = np.array(self.dfc.loc[test_idx, "Energy_Pred"].tolist())
+            ret["test"]["mae"] = np.mean(abs(true_test - pred_test))
+            ret["test"]["rmse"] = np.sqrt(np.mean(np.square(true_test - pred_test)))
+        elif (quantity == "Force"):
+            test_bool = self.dfa.loc[:, "Testing_Bool"].tolist()
+            test_idx = [i for i, x in enumerate(test_bool) if x]
[…]
