### 2023-02-14 6e3d992
Don't print force/energy comparisons in training loop
