### 2023-02-14 e0f270f
Use 32 bit float when evaluating configs
