### 2023-02-15 397fa31
Repeat CI with printed finite difference
