+++ examples/WBe_PyTorch_NN/WBe-example.in
@@ -27,7 +27,7 @@ W = 0.0
 Be = 0.0
 
 [PYTORCH]
-layer_sizes = num_desc 64 64 1
+layer_sizes = num_desc 32 32 1
 learning_rate = 1.5e-4 
 num_epochs = 10
 batch_size = 4
