### 2023-02-15 eec4f43
Revert scraper test size calculation
