### 2023-02-15 dcad79d
Use peratom dataframe when plotting forces
