### 2023-02-15 d04a9a7
Fix scraper roundoff error
