### 2023-02-15 6b28436
Convert models to float64 when evaluating errors
