+++ fitsnap3lib/lib/neural_networks/pytorch.py
@@ -74,7 +74,6 @@ class FitTorch(torch.nn.Module):
             networks[indx].to(self.dtype)
             setattr(self, "network_architecture"+str(indx), networks[indx])
 
-
         self.networks = networks
 
         # now self.state_dict is populated with the attributes declared above 
@@ -117,6 +116,9 @@ class FitTorch(torch.nn.Module):
 
         """
 
+        #print(x.dtype)
+        #for param_tensor in self.state_dict():
+        #    print(param_tensor, "\t", self.state_dict()[param_tensor].dtype)   
         if (self.multi_element_option==1):
             per_atom_energies = self.network_architecture0(x)
    
