### 2023-02-15 c02c59a
Use same columns as linear error metrics
