### 2023-02-15 e18d4c6
Optionalize distribution plots in NN tools
