### 2023-02-15 9e5a1ae
Fix tutorial image loading
