### 2023-02-21 6425716
Use CPU when evaluating configs after training on GPU
