### 2023-02-21 83f40ca
Write model to CPU after training on GPU
