### 2023-02-22 1156635
updating group weights for Fe quadratic example
