### 2023-02-23 #5118 @rohskopf
Hi @tpurcell90 and @bahung, hope all is well.

Feel free to try the Colab tutorial for installing FitSNAP and performing fits:
https://colab.research.google.com/github/FitSNAP/FitSNAP/blob/master/tutorial.ipynb

This tutorial applies what is found in the docs:
https://fitsnap.github.io/
