### 2023-02-25 #5118 @tpurcell90
Comments on the installation:

1. In the quick instructions the load modules section has:

```
module purge
module load gcc
module load cmake
module load openmpi
module load intel # Sometimes this helps mpi4py work
```

I am confused why the intel module would be needed for mpi4py. mpi4py should use whatever MPI library/executable was active when `pip install mpi4py` was run (conda is different and uses a conda-internal MPI library which can complicate things). My concern is that by loading both modules there could be an MPI mismatch between LAMMPS and FitSNAP that can lead to a failure in theory (I have not tested this explicitly)

2. `-DPKG_ML-PACE=yes \` is in the non-PACE install instructions, but you have a separate set of instructions for PACE. Is this intentional?
3. Installing LAMMPS using CMake I get the following warning:

```
CMake Warning at CMakeLists.txt:208 (add_library):
  Cannot generate a safe runtime search path for target lammps because files
  in some directories may conflict with libraries in implicit directories:

    runtime library [libz.so.1] in /usr/lib/x86_64-linux-gnu may be hidden by files in:
      /home/tpurcell/anaconda3/envs/fitsnap/lib
    runtime library [libgomp.so.1] in /usr/lib/gcc/x86_64-linux-gnu/9 may be hidden by files in:
      /home/tpurcell/anaconda3/envs/fitsnap/lib

  Some of these libraries may not be found correctly.

CMake Warning at CMakeLists.txt:215 (add_executable):
  Cannot generate a safe runtime search path for target lmp because files in
  some directories may conflict with libraries in implicit directories:

    runtime library [libgomp.so.1] in /usr/lib/gcc/x86_64-linux-gnu/9 may be hidden by files in:
      /home/tpurcell/anaconda3/envs/fitsnap/lib

  Some of these libraries may not be found correctly.
```

This is a common issue when dealing with conda, I'll see if this causes runtime issues. If it doesn't you may want to add a blurb warning about this issue for a user who may not be as experienced.

4.  Building LAMMPS I see a decent amount of `-Weffc++` and `-Wunused-result` warnings. Depending on who controls the LAMMPS code you might just want to recommend turning off those warnings in the installation instructions. They should not affect the runtime performance of the code.
5. For `make install-python` I would just add a warning that if you did not create a separate virtual env for the code you may need sudo privileges (I ran into this issue in the past)
6. When installing FitSNAP, why isn't there a simple setup.py script that installs the package into the `site-packages` directory? Setting the PYTHONPATH works, but it is easier just to install the code using pip or python setup.py install

(I will edit once I do the run FitSNAP options to see if I encounter any other issues)
