### 2023-02-25 #5118 @tpurcell90
Input File Examples
1. For flags a lot of the values are 0 or 1, why don't you True and False?
2. This maybe more work than it's worth, but for scrapers would it be possible to incorporate something like ASE or pymatgen to read in more types of output files? This could potentially be done with using only their io routines to convert the structures into xyz files. I bring this up since you mention in the docs you will try to use vasp files directly, and this could be an easy workaround to extend it to other codes.
3. I would list out all terms in [MEMORY] and [EXTRAS], I know it is not likely to change, but its good to list them.
4. I would also add some more basic physics/math tutorials to the user guide. The documentation does a good job of describing what each parameter does, but not the underlying physics/mathematical of why those need to change. This is also dependent on the target user if you expect to use the tool.
