### 2023-02-25 #5118 @tpurcell90
`exceptionfitsnap3lib.parallel_tools.GracefulError(*args, **kwargs)` has no docstring. This maybe intentional, but worth noting

Otherwise the library API is good
