### 2023-02-25 #5118 @tpurcell90
Running the tests with intel-MPI fails with the following error:
```
match_arg (../../../../../src/pm/i_hydra/libhydra/arg/hydra_arg.c:91): unrecognized argument oversubscribe
```
I believe that this is an artifact of using openMPI arguments for a non-openMPI library. Should I recompile everything using openMPI?
