### 2023-02-25 #5118 @tpurcell90
In the collab notebook:
For multi-element NNs section: the runs fail with the following traceback
```
/content/FitSNAP/examples/WBe_PyTorch_NN
Traceback (most recent call last):
  File "plot_comparison.py", line 15, in <module>
    nnt = NNTools("peratom.dat", "perconfig.dat")
  File "/content/FitSNAP/fitsnap3lib/tools/nn_tools.py", line 32, in __init__
    self.dfa = pd.read_csv(peratom_file,
  File "/usr/local/lib/python3.8/dist-packages/pandas/util/_decorators.py", line 311, in wrapper
    return func(*args, **kwargs)
  File "/usr/local/lib/python3.8/dist-packages/pandas/io/parsers/readers.py", line 586, in read_csv
    return _read(filepath_or_buffer, kwds)
  File "/usr/local/lib/python3.8/dist-packages/pandas/io/parsers/readers.py", line 482, in _read
    parser = TextFileReader(filepath_or_buffer, **kwds)
  File "/usr/local/lib/python3.8/dist-packages/pandas/io/parsers/readers.py", line 811, in __init__
    self._engine = self._make_engine(self.engine)
  File "/usr/local/lib/python3.8/dist-packages/pandas/io/parsers/readers.py", line 1040, in _make_engine
    return mapping[engine](self.f, **self.options)  # type: ignore[call-arg]
  File "/usr/local/lib/python3.8/dist-packages/pandas/io/parsers/c_parser_wrapper.py", line 51, in __init__
    self._open_handles(src, kwds)
  File "/usr/local/lib/python3.8/dist-packages/pandas/io/parsers/base_parser.py", line 222, in _open_handles
    self.handles = get_handle(
  File "/usr/local/lib/python3.8/dist-packages/pandas/io/common.py", line 702, in get_handle
    handle = open(
FileNotFoundError: [Errno 2] No such file or directory: 'peratom.dat'
```

Run MD with WBe NN-SNAP fails with:

```
ERROR on proc 0: Cannot open SNAP parameter file ../WBe_pot.mliap.descriptor: No such file or directory (src/ML-IAP/mliap_descriptor_snap.cpp:379)
Last command: pair_style hybrid/overlay zbl 4.0 4.8 mliap model mliappy ../FitTorch_Pytorch.pt descriptor sna ../WBe_pot.mliap.descriptor
--------------------------------------------------------------------------
MPI_ABORT was invoked on rank 0 in communicator MPI_COMM_WORLD
with errorcode 1.
```

Pairwise netoworks fails with:
```%cd /content/FitSNAP/examples/Ta_pytorch_custom_NN/

!python -m fitsnap3 Ta-example.in --overwrite

Average loss over batches is 0.006325966378100789
'decorated_perform_fit' took 1518653.77 ms on rank 0
'fit_gather' took 0.01 ms on rank 0
Traceback (most recent call last):
  File "/content/FitSNAP/fitsnap3/__main__.py", line 48, in main
    snap.perform_fit()
  File "/content/FitSNAP/fitsnap3lib/fitsnap.py", line 152, in perform_fit
    error_analysis()
  File "/content/FitSNAP/fitsnap3lib/parallel_tools.py", line 277, in timed
    result = method(*args, **kw)
  File "/content/FitSNAP/fitsnap3lib/fitsnap.py", line 148, in error_analysis
    self.solver.error_analysis()
  File "/content/FitSNAP/fitsnap3lib/solvers/solver.py", line 338, in error_analysis
    decorated_error_analysis()
  File "/content/FitSNAP/fitsnap3lib/parallel_tools.py", line 308, in check_if_rank_zero
    return method(*args, **kw)
  File "/content/FitSNAP/fitsnap3lib/solvers/solver.py", line 184, in decorated_error_analysis
    mae_e[c.group]["train"] += ae
KeyError: ''

Traceback (most recent call last):
  File "/usr/lib/python3.8/runpy.py", line 194, in _run_module_as_main
    return _run_code(code, main_globals, None,
  File "/usr/lib/python3.8/runpy.py", line 87, in _run_code
    exec(code, run_globals)
  File "/content/FitSNAP/fitsnap3/__main__.py", line 55, in <module>
    main()
  File "/content/FitSNAP/fitsnap3lib/parallel_tools.py", line 277, in timed
    result = method(*args, **kw)
  File "/content/FitSNAP/fitsnap3/__main__.py", line 51, in main
    output.exception(e)
TypeError: exception() missing 1 required positional argument: 'err'
```
