### 2023-02-25 #5118 @rohskopf
Thank you @tpurcell90 for the  detailed review. I will start  addressing your points.

Feel free to open issues on our repo as well.
