### 2023-02-25 #5118 @jmschrei
Thank you for the review, @tpurcell90. There is no requirement to open issues in the repo, but I've found that opening issues can be useful for tracking the issue if the fix ends up being more complicated than expected, and allows others to comment on it if they encounter a similar problem in the future.
