### 2023-02-25 07c336e
Dump per-atom data in WBe NN example
