### 2023-02-25 f987d40
Make output class for custom/pair NNs
