+++ fitsnap3lib/io/outputs/output_factory.py
@@ -2,6 +2,7 @@ from fitsnap3lib.parallel_tools import ParallelTools
 from fitsnap3lib.io.outputs.outputs import Output
 from fitsnap3lib.io.outputs.pace import Pace
 from fitsnap3lib.io.outputs.snap import Snap
+from fitsnap3lib.io.outputs.custom import Custom
 
 
 #pt = ParallelTools()
