+++ examples/Ta_pytorch_custom_NN/Ta-example.in
@@ -19,9 +19,9 @@ nonlinear = 1
 Ta = 0.0
 
 [NETWORK]
-layer_sizes = num_desc 32 16 1 
+layer_sizes = num_desc 32 32 1 
 learning_rate = 5e-4
-num_epochs = 150
+num_epochs = 10 # Run for ~150 epochs to get < 100 meV/A error
 batch_size = 4 # 363 configs in entire set
 save_state_output = Ta_Pytorch.pt
 energy_weight = 1e-2
@@ -41,6 +41,7 @@ scraper = JSON
 dataPath = ../Ta_Linear_JCP2014/JSON
 
 [OUTFILE]
+output_style = custom
 metrics = Ta_metrics.md
 potential = Ta_pot
 
@@ -70,9 +71,8 @@ Volume_BCC    =  0.7    0.3      1.00E+00        1.00E-09        1.00E-09
 Volume_FCC    =  0.7    0.3      1.00E+00        1.00E-09        1.00E-09
 
 [EXTRAS]
-dump_descriptors = 1
-dump_truth = 1
-dump_weights = 1
+dump_peratom = 1
+dump_perconfig = 1
 
 [MEMORY]
 override = 0
