### 2023-02-26 FitSNAP/FitSNAP#180 @tpurcell90
JOSS Review: Adding a setup.py script to install FitSNAP
With this one would be able to have the package in their `site-packages` directory without the need to reset the `PYTHONPATH`. This can be useful if you run multiple virtual envs with different versions.

It should be relatively straight forward to do, here is a link for documentation https://docs.python.org/3/distutils/setupscript.html
