### 2023-02-26 FitSNAP/FitSNAP#181 @tpurcell90
JOSS Review: Add Boolean Flags for input files
Instead of using 1 and 0 for the flags you can use True/False with configparser's getboolean function: https://docs.python.org/3/library/configparser.html#configparser.ConfigParser.getboolean

This can make the inputs more readable so if something is 0 then it means a numeric 0 and not a flag option
