### 2023-02-26 FitSNAP/FitSNAP#182 @tpurcell90
JOSS Review: Linting to Enforce coding standards
Using a linter can help ensure coding standards you mention in the guidelines, and spot some errors before running. I personally use black, but there are multiple types with various pros/cons. Here is a link that describes many of

them:https://inventwithpython.com/blog/2022/11/19/python-linter-comparison-2022-pylint-vs-pyflakes-vs-flake8-vs-autopep8-vs-bandit-vs-prospector-vs-pylama-vs-pyroma-vs-black-vs-mypy-vs-radon-vs-mccabe/
