### 2023-02-26 FitSNAP/FitSNAP#183 @tpurcell90
JOSS Review: Tests Fail when Using Intel MPI (OneAPI version 2021.6.0)
When running `test_examples.py` I get the following errors

```
.[mpiexec@nomadbook-7] match_arg (../../../../../src/pm/i_hydra/libhydra/arg/hydra_arg.c:91): unrecognized argument oversubscribe
[mpiexec@nomadbook-7] HYD_arg_parse_array (../../../../../src/pm/i_hydra/libhydra/arg/hydra_arg.c:128): argument matching returned error
[mpiexec@nomadbook-7] mpiexec_get_parameters (../../../../../src/pm/i_hydra/mpiexec/mpiexec_params.c:1359): error parsing input array
[mpiexec@nomadbook-7] main (../../../../../src/pm/i_hydra/mpiexec/mpiexec.c:1783): error parsing parameters
```

This likely is using an OpenMPI only argument. I'm not sure how easy it is to fix or if one can just use only OpenMPI for the code.
