### 2023-02-26 FitSNAP/FitSNAP#185 @tpurcell90
JOSS Review: Warnings when bulding LAMMPS
When building LAMMPS following the instructions there are multiple `-Weffc++` and `-Wunused-result` warnings in the compilation. While not great they should not affect the results and unless you are on the LAMMPS team and can easily fix them I'd just set `-Wno-effc++` and `-Wno-unused-result` in the CXXFLAGS. Or add a warning in the documentation about it.
