### 2023-02-26 FitSNAP/FitSNAP#186 @tpurcell90
JOSS Review Optional: ASE/pymatgen Scrappers
A quality of life solution could use existing libraries to pull geometry files directly from electronic structure codes. This is completely optional since adding dependencies can make things more complicated to manage.
