### 2023-02-26 #5118 @tpurcell90
I added 5 issues and one optional issue for the code recommendations/actual errors I found all headed with JOSS Review prefix. The other comments aren't really issues and more questions so I did not add an issue onto the repo.
