### 2023-02-26 abbeb1d
Make examples more tractable for tutorial
