+++ setup.cfg
@@ -1,0 +1,25 @@
+[metadata]
+name = fitsnap3
+version = 3.0.0
+author = Drew Rohskopf
+author_email = drew.rohskopf@gmail.com
+description = Interatomic Potential Machine Learning Interface
+long_description = file: README.md, LICENSE
+long_description_content_type = text/markdown
+url = https://github.com/rohskopf/FitSNAP/tree/package
+project_urls =
+    Bug Tracker = https://github.com/rohskopf/FitSNAP/-/issues
+    repository = https://github.com/rohskopf/FitSNAP/tree/package
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: GNU General Public License (GPL)
+    Operating System :: OS Independent
+
+[options]
+package_dir =
+    = fitsnap3lib
+packages = find:
+python_requires = >=3.6
+
+[options.packages.find]
+where = fitsnap3lib
