### 2023-02-27 ef395d6
Update setup cfg
