+++ setup.cfg
@@ -1,9 +1,9 @@
 [metadata]
 name = fitsnap3
-version = 3.0.0
+version = 3.0.1
 author = Drew Rohskopf
 author_email = drew.rohskopf@gmail.com
-description = Interatomic Potential Machine Learning Interface
+description = Molecular Machine Learning Interface
 long_description = file: README.md, LICENSE
 long_description_content_type = text/markdown
 url = https://github.com/rohskopf/FitSNAP/tree/package
@@ -12,14 +12,9 @@ project_urls =
     repository = https://github.com/rohskopf/FitSNAP/tree/package
 classifiers =
     Programming Language :: Python :: 3
-    License :: OSI Approved :: GNU General Public License (GPL)
+    License :: OSI Approved :: MIT License
     Operating System :: OS Independent
 
 [options]
-package_dir =
-    = fitsnap3lib
-packages = find:
+packages = fitsnap3, fitsnap3lib
 python_requires = >=3.6
-
-[options.packages.find]
-where = fitsnap3lib
