### 2023-02-27 d84e499
Confirm setup.cfg works with local pip install
