+++ setup.cfg
@@ -2,13 +2,13 @@
 name = fitsnap3
 version = 3.0.1
 author = Drew Rohskopf
-author_email = drew.rohskopf@gmail.com
+author_email = adrohsk@sandia.gov
 description = Molecular Machine Learning Interface
 long_description = file: README.md, LICENSE
 long_description_content_type = text/markdown
 url = https://github.com/rohskopf/FitSNAP/tree/package
 project_urls =
-    Bug Tracker = https://github.com/rohskopf/FitSNAP/-/issues
+    Bug Tracker = https://github.com/FitSNAP/FitSNAP/-/issues
     repository = https://github.com/rohskopf/FitSNAP/tree/package
 classifiers =
     Programming Language :: Python :: 3
