### 2023-02-27 9c5683a
Remove eggs
