+++ fitsnap3lib/fitsnap3.egg-info/SOURCES.txt
@@ -1,107 +1,0 @@
-LICENSE
-README.md
-pyproject.toml
-setup.cfg
-fitsnap3lib/calculators/__init__.py
-fitsnap3lib/calculators/basic_calculator.py
-fitsnap3lib/calculators/calculator.py
-fitsnap3lib/calculators/calculator_factory.py
-fitsnap3lib/calculators/lammps_base.py
-fitsnap3lib/calculators/lammps_custom.py
-fitsnap3lib/calculators/lammps_pace.py
-fitsnap3lib/calculators/lammps_snap.py
-fitsnap3lib/calculators/template_calculator.py
-fitsnap3lib/fitsnap3.egg-info/PKG-INFO
-fitsnap3lib/fitsnap3.egg-info/SOURCES.txt
-fitsnap3lib/fitsnap3.egg-info/dependency_links.txt
-fitsnap3lib/fitsnap3.egg-info/top_level.txt
-fitsnap3lib/io/__init__.py
-fitsnap3lib/io/error.py
-fitsnap3lib/io/input.py
-fitsnap3lib/io/output.py
-fitsnap3lib/io/outputs/__init__.py
-fitsnap3lib/io/outputs/output_factory.py
-fitsnap3lib/io/outputs/outputs.py
-fitsnap3lib/io/outputs/pace.py
-fitsnap3lib/io/outputs/snap.py
-fitsnap3lib/io/outputs/template_output.py
-fitsnap3lib/io/sections/__init__.py
-fitsnap3lib/io/sections/eshift.py
-fitsnap3lib/io/sections/extras.py
-fitsnap3lib/io/sections/groups.py
-fitsnap3lib/io/sections/memory.py
-fitsnap3lib/io/sections/outfile.py
-fitsnap3lib/io/sections/path.py
-fitsnap3lib/io/sections/reference.py
-fitsnap3lib/io/sections/scraper.py
-fitsnap3lib/io/sections/section_factory.py
-fitsnap3lib/io/sections/sections.py
-fitsnap3lib/io/sections/template.py
-fitsnap3lib/io/sections/trainshift.py
-fitsnap3lib/io/sections/calculator_sections/__init__.py
-fitsnap3lib/io/sections/calculator_sections/ace.py
-fitsnap3lib/io/sections/calculator_sections/basic_calculator.py
-fitsnap3lib/io/sections/calculator_sections/bispectrum.py
-fitsnap3lib/io/sections/calculator_sections/calculator.py
-fitsnap3lib/io/sections/calculator_sections/custom.py
-fitsnap3lib/io/sections/solver_sections/__init__.py
-fitsnap3lib/io/sections/solver_sections/ard.py
-fitsnap3lib/io/sections/solver_sections/jax.py
-fitsnap3lib/io/sections/solver_sections/lasso.py
-fitsnap3lib/io/sections/solver_sections/network.py
-fitsnap3lib/io/sections/solver_sections/pytorch.py
-fitsnap3lib/io/sections/solver_sections/solver.py
-fitsnap3lib/lib/__init__.py
-fitsnap3lib/lib/neural_networks/__init__.py
-fitsnap3lib/lib/neural_networks/jax.py
-fitsnap3lib/lib/neural_networks/pairwise.py
-fitsnap3lib/lib/neural_networks/pas.py
-fitsnap3lib/lib/neural_networks/pytorch.py
-fitsnap3lib/lib/neural_networks/write.py
-fitsnap3lib/lib/scalapack_solver/__init__.py
-fitsnap3lib/lib/scalapack_solver/scalapack.py
[…]
