+++ fitsnap3lib/fitsnap3.egg-info/top_level.txt
@@ -1,7 +1,0 @@
-calculators
-io
-lib
-scrapers
-solvers
-tools
-units
