+++ setup.cfg
@@ -6,10 +6,10 @@ author_email = adrohsk@sandia.gov
 description = Molecular Machine Learning Interface
 long_description = file: README.md, LICENSE
 long_description_content_type = text/markdown
-url = https://github.com/rohskopf/FitSNAP/tree/package
+url = https://github.com/FitSNAP/FitSNAP
 project_urls =
     Bug Tracker = https://github.com/FitSNAP/FitSNAP/-/issues
-    repository = https://github.com/rohskopf/FitSNAP/tree/package
+    repository = https://github.com/FitSNAP/FitSNAP
 classifiers =
     Programming Language :: Python :: 3
     License :: OSI Approved :: MIT License
