+++ pyproject.toml
@@ -1,3 +1,3 @@
 [build-system]
-requires = ['setuptools>=42']
+requires = ['setuptools>=42', 'numpy', 'scipy', 'scikit-learn', 'virtualenv', 'psutil', 'pandas', 'tabulate', 'mpi4py', 'Cython', 'torch', 'sympy', 'pyyaml']
 build-backend = 'setuptools.build_meta'
