+++ setup.cfg
@@ -1,20 +1,0 @@
-[metadata]
-name = fitsnap3
-version = 3.0.1
-author = Drew Rohskopf
-author_email = adrohsk@sandia.gov
-description = Molecular Machine Learning Interface
-long_description = file: README.md, LICENSE
-long_description_content_type = text/markdown
-url = https://github.com/FitSNAP/FitSNAP
-project_urls =
-    Bug Tracker = https://github.com/FitSNAP/FitSNAP/-/issues
-    repository = https://github.com/FitSNAP/FitSNAP
-classifiers =
-    Programming Language :: Python :: 3
-    License :: OSI Approved :: MIT License
-    Operating System :: OS Independent
-
-[options]
-packages = fitsnap3, fitsnap3lib
-python_requires = >=3.6
