### 2023-02-28 FitSNAP/FitSNAP#180 @rohskopf
Hi @tpurcell90 , I added a `pyproject.toml` which works nicely with `pip`. Do you have any opinions on `pyproject.toml` vs. `setup.py`?
