### 2023-02-28 FitSNAP/FitSNAP#180 @tpurcell90
They are equivalent in my opinion. setup.py is more common, but pyproject also works.
