### 2023-02-28 #5118 @tpurcell90
I have no opinion. I think `setup.py` in has more general capabilities, but they both work in this case.

I'll check the tutorials later this week.
