### 2023-02-28 FitSNAP/FitSNAP#184 @rohskopf
Didn't mean to close this. Will leave it open while we're addressing other changes that might affect the tutorial.
