### 2023-02-28 #5118 @jmschrei
I don't have a strong opinion either way. I personally use `setup.py` but that's because that was the trend when I learned how to do these things.
