### 2023-02-28 e8c26ad
Loop over individual config evaluations when calculating errors
