+++ fitsnap3lib/solvers/pytorch.py
@@ -580,11 +580,12 @@ try:
             decorated_perform_fit()
 
         #@pt.sub_rank_zero
-        def evaluate_configs(self, option = 1, standardize_bool = True, dtype=torch.float64, eval_device='cpu'):
+        def evaluate_configs(self, config_idx = 0, option = 1, standardize_bool = True, dtype=torch.float64, eval_device='cpu'):
             """
             Evaluates energies and forces on configs for testing purposes. 
 
             Args:
+                config_idx (int): Index of config to evaluate. None if evaluating all configs.
                 option (int): 1 to evaluate energies/forces for all configs separately, 2 for using 
                     dataloader/batch procedure. Currently only 1 is implemented.
                 standardize_bool (bool): True to standardize weights, False otherwise. Useful if 
@@ -636,9 +637,45 @@ try:
 
                 self.model.eval()
 
-                # for evaluating on single config:
+                # Evaluate a single config:
 
-                if (option==1):
+                if (config_idx is not None):
+                    config = self.configs[config_idx]
+                    descriptors = torch.tensor(config.descriptors).requires_grad_(True)
+                    atom_types = torch.tensor(config.types).long()
+                    target = torch.tensor(config.energy).reshape(-1)
+                    # indexing 0th axis with None reshapes the tensor to be 2D for stacking later
+                    weights = torch.tensor(config.weights[None,:])
+                    target_forces = torch.tensor(config.forces)
+                    num_atoms = torch.tensor(config.natoms)
+                    dgrad = torch.tensor(config.dgrad)
+                    dbdrindx = torch.tensor(config.dgrad_indices).long()
+
+                    # convert quantities to desired dtype
+              
+                    descriptors = descriptors.to(dtype)
+                    target = target.to(dtype)
+                    weights = weights.to(dtype)
+                    target_forces = target_forces.to(dtype)
+                    dgrad = dgrad.to(dtype)
+
+                    # make indices upon which to contract per-atom energies for this config
+
+                    config_indices = torch.arange(1).long() # this usually has len(batch) as arg in dataloader
+                    indices = torch.repeat_interleave(config_indices, num_atoms)
+                    
+                    # illustrate what unique_i and unique_j are
+
+                    unique_i = dbdrindx[:,0]
+                    unique_j = dbdrindx[:,1]
+
+                    (energies,forces) = self.model(descriptors, dgrad, indices, num_atoms, 
+                                                  atom_types, dbdrindx, unique_j, unique_i, 
+                                                  eval_device, dtype)
+
+                # Evaluate all configs:
+
+                else:
 
                     energies_configs = []
                     forces_configs = []
@@ -678,38 +715,7 @@ try:
                         energies_configs.append(energies)
                         forces_configs.append(forces)
 
-                    return(energies_configs, forces_configs)
-
-                else:
-                    print("Other options not implemented yet, see example below")
-
-                    """
-                    for epoch in range(self.config.sections["PYTORCH"].num_epochs):
-                        print(f"----- epoch: {epoch}")
-                        start = time()
-
-                        # loop over training data
-
-                        train_losses_step = []
-                        loss = None
-                        self.model.eval() # don't calculate gradients for simple evaluating
-                        for i, batch in enumerate(self.training_loader):
-                            descriptors = batch['x'].to(self.device).requires_grad_(True)
-                            atom_types = batch['t'].to(self.device)
-                            targets = batch['y'].to(self.device) #.requires_grad_(True)
-                            target_forces = batch['y_forces'].to(self.device) #.requires_grad_(True)
-                            indices = batch['i'].to(self.device)
-                            num_atoms = batch['noa'].to(self.device)
[…]
