### 2023-03-01 f49f4b2
Update tests with new config evaluator
