+++ tests/test_pytorch.py
@@ -58,7 +58,8 @@ def test_fd_single_elem():
     snap.process_configs()
     pt.all_barrier()
     snap.solver.create_datasets()
-    (energies_model, forces_model) = snap.solver.evaluate_configs(option=1, standardize_bool=True)
+    #(energies_model, forces_model) = snap.solver.evaluate_configs(option=1, standardize_bool=True)
+    (energies_model, forces_model) = snap.solver.evaluate_configs(config_idx=None, standardize_bool=True)
 
     print(f"Length of data: {len(snap.data)}")
 
@@ -83,7 +84,8 @@ def test_fd_single_elem():
                   snap.calculator.shared_index_dgrad = 0
                   snap.process_configs()
                   snap.solver.create_datasets()
-                  (energies1, forces1) = snap.solver.evaluate_configs(option=1, standardize_bool=False)
+                  #(energies1, forces1) = snap.solver.evaluate_configs(option=1, standardize_bool=False)
+                  (energies1, forces1) = snap.solver.evaluate_configs(config_idx=None, standardize_bool=False)
 
                   # calculate model energy with -h (energy2)
 
@@ -95,7 +97,8 @@ def test_fd_single_elem():
                   snap.calculator.shared_index_dgrad = 0
                   snap.process_configs()
                   snap.solver.create_datasets()
-                  (energies2, forces2) = snap.solver.evaluate_configs(option=1, standardize_bool=False)
+                  #(energies2, forces2) = snap.solver.evaluate_configs(option=1, standardize_bool=False)
+                  (energies2, forces2) = snap.solver.evaluate_configs(config_idx=None, standardize_bool=False)
 
                   # calculate and compare finite difference force
 
@@ -165,7 +168,8 @@ def test_fd_multi_elem():
     snap.process_configs()
     pt.all_barrier()
     snap.solver.create_datasets()
-    (energies_model, forces_model) = snap.solver.evaluate_configs(option=1, standardize_bool=True)
+    #(energies_model, forces_model) = snap.solver.evaluate_configs(config_idx=None, standardize_bool=True)
+    (energies_model, forces_model) = snap.solver.evaluate_configs(config_idx=None, standardize_bool=True)
 
     print(f"Length of data: {len(snap.data)}")
 
@@ -189,7 +193,8 @@ def test_fd_multi_elem():
                   snap.calculator.shared_index_dgrad = 0
                   snap.process_configs()
                   snap.solver.create_datasets()
-                  (energies1, forces1) = snap.solver.evaluate_configs(option=1, standardize_bool=False)
+                  #(energies1, forces1) = snap.solver.evaluate_configs(option=1, standardize_bool=False)
+                  (energies1, forces1) = snap.solver.evaluate_configs(config_idx=None, standardize_bool=False)
 
                   # calculate model energy with -h (energy2)
 
@@ -201,7 +206,8 @@ def test_fd_multi_elem():
                   snap.calculator.shared_index_dgrad = 0
                   snap.process_configs()
                   snap.solver.create_datasets()
-                  (energies2, forces2) = snap.solver.evaluate_configs(option=1, standardize_bool=False)
+                  #(energies2, forces2) = snap.solver.evaluate_configs(config_idx=None, standardize_bool=False)
+                  (energies2, forces2) = snap.solver.evaluate_configs(config_idx=None, standardize_bool=False)
 
                   # calculate and compare finite difference force
 
@@ -275,7 +281,7 @@ def test_fd_ace_single_elem():
     snap.process_configs()
     pt.all_barrier()
     snap.solver.create_datasets()
-    (energies_model, forces_model) = snap.solver.evaluate_configs(option=1, standardize_bool=True)
[…]
