+++ fitsnap3lib/solvers/solver.py
@@ -154,17 +154,6 @@ class Solver:
                 count_train['*ALL']["nconfigs"] = 0 # Total number test configs in group.
                 count_train['*ALL']["natoms"] = 0 # Total number test atoms in group.
 
-                # Evaluate errors with float64 dtype since this is what we use in production.
-
-                if (self.config.sections["CALCULATOR"].calculator == "LAMMPSCUSTOM"):
-                    #(energies_model, forces_model) = self.evaluate_configs(option=1, evaluate_all = True, standardize_bool=False, dtype=torch.float64)
-                    pass
-                else:
-                    #(energies_model, forces_model) = self.evaluate_configs(config_idx=0, option=1, standardize_bool=False, dtype=torch.float64)
-                    #print(energies_model)
-                    #print(forces_model)
-                    #assert(False)
-                    pass
                 if (self.config.sections["EXTRAS"].dump_peratom):
                     fha = open(self.config.sections["EXTRAS"].peratom_file, 'w')
                     line = f"Filename Group AtomID Type Fx_Truth Fy_Truth Fz_Truth Fx_Pred Fy_Pred Fz_Pred Testing_Bool"
