### 2023-03-02 FitSNAP/FitSNAP#182 @rohskopf
Thanks for the link. Currently we have flake8 in our CI tests but it's just raising warnings. I'll put it in a setup.cfg.
