### 2023-03-02 FitSNAP/FitSNAP#184 @rohskopf
These errors are fixed. Just small changes to abide by the new NN outputs introduced in https://github.com/FitSNAP/FitSNAP/pull/174
