### 2023-03-02 FitSNAP/FitSNAP#185 @rohskopf
Most of these warnings come from the PACE package for ACE descriptors. This and the ML-IAP package are under a lot of development, so these warnings are useful sometimes. It's best to note these warnings in the docs so that users aren't alarmed.
