### 2023-03-02 FitSNAP/FitSNAP#186 @rohskopf
Definitely a good quality of life solution. I'm hesitant to add more dependencies so this is nice to have as an option, like you suggested. We therefore made a library example that loads data in XYZ files, then inputs them into FitSNAP:
https://github.com/FitSNAP/FitSNAP/blob/master/examples/library/single_config/example.py

I will note this example in the docs, and elaborate on the fact that ASE can be used to scrape more than just XYZ, but also VASP and other output.
