### 2023-03-02 FitSNAP/FitSNAP#183 @rohskopf
Thanks for pointing this out. We never noticed this since our CI tests use OpenMPI. I think it's using an OpenMPI only argument, I will clarify. If OpenMPI is absolutely required then that should be documented.
