### 2023-03-03 FitSNAP/FitSNAP#182 @tpurcell90
okay, some code bases also explicitly throw an error for linting errors, I am not sure how strict one wants to be about that.
