### 2023-03-03 FitSNAP/FitSNAP#183 @tpurcell90
I saw a workaround when looking up the error on stack overflow, but I don't know how easy it will be to implement
