### 2023-03-03 FitSNAP/FitSNAP#185 @tpurcell90
Okay that makes sense, the warnings don't affect the code from what I could tell so a note in the docs should also work
