### 2023-03-03 FitSNAP/FitSNAP#186 @tpurcell90
That is a good middle ground
