+++ fitsnap3lib/lib/sym_ACE/lib/sylow_lib.py
@@ -328,86 +328,86 @@ fusion8 = [((4, 6, 5, 7),),
 
 
 def filled_perm(tups,rank):
-	allinds = list(range(rank))
-	try:
-		remainders = [ i for i in allinds if i not in flatten(tups)]
-		alltups = tups + tuple([tuple([k]) for k in remainders])
-	except TypeError:
-		remainders = [ i for i in allinds if i not in flatten(flatten(tups))]
-		alltups = tups + tuple([tuple([k]) for k in remainders])
-	return alltups
+    allinds = list(range(rank))
+    try:
+        remainders = [ i for i in allinds if i not in flatten(tups)]
+        alltups = tups + tuple([tuple([k]) for k in remainders])
+    except TypeError:
+        remainders = [ i for i in allinds if i not in flatten(flatten(tups))]
+        alltups = tups + tuple([tuple([k]) for k in remainders])
+    return alltups
 
 def base_automorphisms(l,subtree=False):
-	# base automorphism groups for most practical descriptor labels.
-	if len(l) == 4 or len(l) == 5:
-		if not subtree:
-			all_ops = sylow_2_s4
-		else:
-			all_ops = s4_subtree_syms
-	elif len(l) == 6 or len(l) == 7:
-		if not subtree:
-			all_ops = sylow_2_s6
-		else:
-			all_ops = s6_subtree_syms
-	elif len(l) == 8 or len(l) == 9:
-		if not subtree:
-			all_ops = sylow_2_s8
-		else:
-			all_ops = s8_subtree_syms
-	else:
-		raise ValueError()
+    # base automorphism groups for most practical descriptor labels.
+    if len(l) == 4 or len(l) == 5:
+        if not subtree:
+            all_ops = sylow_2_s4
+        else:
+            all_ops = s4_subtree_syms
+    elif len(l) == 6 or len(l) == 7:
+        if not subtree:
+            all_ops = sylow_2_s6
+        else:
+            all_ops = s6_subtree_syms
+    elif len(l) == 8 or len(l) == 9:
+        if not subtree:
+            all_ops = sylow_2_s8
+        else:
+            all_ops = s8_subtree_syms
+    else:
+        raise ValueError()
 
-	all_ops.sort(key=lambda x: x,reverse=True)
-	all_ops.sort(key=lambda x: tuple([len(k) for k in x]),reverse=True)
+    all_ops.sort(key=lambda x: x,reverse=True)
+    all_ops.sort(key=lambda x: tuple([len(k) for k in x]),reverse=True)
 
-	all_ops = [filled_perm(op,len(l)) for op in all_ops]
+    all_ops = [filled_perm(op,len(l)) for op in all_ops]
 
-	all_parts = []
-	for op in all_ops:
-		orbits = [len(x) for x in op]
-		partition = tuple(orbits)
-		all_parts.append(partition)
-	unique_parts = list(set(all_parts))
-	unique_parts.sort(key=lambda x: len(x),reverse=True)
-	unique_parts.sort(key=lambda x: x,reverse=True)
-	op_per_part = {p:[] for p in unique_parts}
-	for op_part in zip(all_ops,all_parts):
-		this_op,this_part = op_part
-		op_per_part[this_part].append(this_op)
+    all_parts = []
+    for op in all_ops:
+        orbits = [len(x) for x in op]
+        partition = tuple(orbits)
+        all_parts.append(partition)
+    unique_parts = list(set(all_parts))
+    unique_parts.sort(key=lambda x: len(x),reverse=True)
+    unique_parts.sort(key=lambda x: x,reverse=True)
+    op_per_part = {p:[] for p in unique_parts}
+    for op_part in zip(all_ops,all_parts):
+        this_op,this_part = op_part
+        op_per_part[this_part].append(this_op)
 
[…]
