+++ fitsnap3lib/lib/sym_ACE/coupling_coeffs.py
@@ -5,104 +5,104 @@ from fitsnap3lib.lib.sym_ACE.sym_ACE_settings import *
 
 
 def Clebsch_gordan(j1,m1,j2,m2,j3,m3):
-	# Clebsch-gordan coefficient calculator based on eqs. 4-5 of:
-	# https://hal.inria.fr/hal-01851097/document
-	# and christoph ortner's julia code ACE.jl
+    # Clebsch-gordan coefficient calculator based on eqs. 4-5 of:
+    # https://hal.inria.fr/hal-01851097/document
+    # and christoph ortner's julia code ACE.jl
 
-	#VERIFIED: test non-zero indices in Wolfram using format ClebschGordan[{j1,m1},{j2,m2},{j3,m3}]
-	#rules:
-	rule1 = np.abs(j1-j2) <= j3
-	rule2 = j3 <= j1+j2
-	rule3 = m3 == m1 + m2
-	rule4 = np.abs(m3) <= j3
+    #VERIFIED: test non-zero indices in Wolfram using format ClebschGordan[{j1,m1},{j2,m2},{j3,m3}]
+    #rules:
+    rule1 = np.abs(j1-j2) <= j3
+    rule2 = j3 <= j1+j2
+    rule3 = m3 == m1 + m2
+    rule4 = np.abs(m3) <= j3
 
-	#rules assumed by input
-	#assert np.abs(m1) <= j1, 'm1 must be \in {-j1,j1}'
-	#assert np.abs(m2) <= j2, 'm2 must be \in {-j2,j2}'
+    #rules assumed by input
+    #assert np.abs(m1) <= j1, 'm1 must be \in {-j1,j1}'
+    #assert np.abs(m2) <= j2, 'm2 must be \in {-j2,j2}'
 
-	if rule1 and rule2 and rule3 and rule4:
-		#attempting binomial representation
-		N1 = np.longdouble((2*j3) + 1 )
-		N2 =  np.longdouble(special.factorial(j1 + m1, exact=True)) \
-		* np.longdouble(special.factorial(j1 - m1, exact=True)) \
-		* np.longdouble(special.factorial(j2 + m2, exact=True)) \
-		* np.longdouble(special.factorial(j2 - m2, exact=True)) \
-		* np.longdouble(special.factorial(j3 + m3, exact=True)) \
-		* np.longdouble(special.factorial(j3 - m3, exact=True))
+    if rule1 and rule2 and rule3 and rule4:
+        #attempting binomial representation
+        N1 = np.longdouble((2*j3) + 1 )
+        N2 =  np.longdouble(special.factorial(j1 + m1, exact=True)) \
+        * np.longdouble(special.factorial(j1 - m1, exact=True)) \
+        * np.longdouble(special.factorial(j2 + m2, exact=True)) \
+        * np.longdouble(special.factorial(j2 - m2, exact=True)) \
+        * np.longdouble(special.factorial(j3 + m3, exact=True)) \
+        * np.longdouble(special.factorial(j3 - m3, exact=True))
 
-		N3 = np.longdouble(special.factorial(j1 + j2 - j3, exact=True)) \
-		* np.longdouble(special.factorial(j1 - j2 + j3, exact=True)) \
-		* np.longdouble(special.factorial(-j1 + j2 + j3, exact=True)) \
-		* np.longdouble(special.factorial(j1 + j2 + j3 + 1, exact=True))
+        N3 = np.longdouble(special.factorial(j1 + j2 - j3, exact=True)) \
+        * np.longdouble(special.factorial(j1 - j2 + j3, exact=True)) \
[…]
