### 2023-03-03 FitSNAP/FitSNAP#181 @rohskopf
Thanks for pointing this out.

We have a custom `get_value` function that allows `True` or `False` along with `1` or `0` in the config file:
https://github.com/FitSNAP/FitSNAP/blob/c6cb10ea67e4d73d172c7f1229c812da9b67cd1d/fitsnap3lib/io/sections/sections.py#L52

This allows us to use the same function for all inputs whether users prefer using Booleans or ones/zeros. For example choosing `force = 1` or `force = True` in the `[CALCULATOR]` section will turn on force fitting:
https://github.com/FitSNAP/FitSNAP/blob/c6cb10ea67e4d73d172c7f1229c812da9b67cd1d/fitsnap3lib/io/sections/calculator_sections/calculator.py#L26

Sometimes we have flags that take on values beyond 1 and 0, and we can still use the same `get_value` function for these.

Overall we prefer keeping this function the way it is because it is used in all of our inputs for various purposes. In the end, all Boolean flags get converted to the proper data type internally.
