### 2023-03-03 d271865
Add flake8 in setup.cfg for #182
