+++ setup.cfg
@@ -1,0 +1,13 @@
+[flake8]
+# Invoke this flake8 setup by calling `flake8` in the terminal.
+max-line-length = 127
+select = E123,E101,W191,E111,E114,E116,E117,F811,F401,W391
+exclude =
+    .git,
+    __pycache__,
+    docs,
+    build,
+    dist,
+    .eggs,
+    *.egg
+max-complexity = 10
