+++ tests/example_checker.py
@@ -11,252 +11,252 @@ from runpy import run_module
 
 
 class ExampleChecker:
-  """ Used to check example output from pytest tests """
+    """ Used to check example output from pytest tests """
 
-  def __init__(self, input_file):
-    self._input_file = input_file
-    self._example_path = input_file.parent
-    self._config = configparser.ConfigParser(inline_comment_prefixes='#')
-    self._config.optionxform = str
-    self._config.read(str(input_file))
-    self._outfile = self._config.get("OUTFILE", "potential", fallback="fitsnap_potential")
-    self._standard_path = self._check_for_standard()
-    self._mpi = MPICheck()
+    def __init__(self, input_file):
+        self._input_file = input_file
+        self._example_path = input_file.parent
+        self._config = configparser.ConfigParser(inline_comment_prefixes='#')
+        self._config.optionxform = str
+        self._config.read(str(input_file))
+        self._outfile = self._config.get("OUTFILE", "potential", fallback="fitsnap_potential")
+        self._standard_path = self._check_for_standard()
+        self._mpi = MPICheck()
 
-  def _check_for_standard(self):
-    """ Determine directory that contains the output standard """
-    possibilites = self._example_path.glob('*_Standard')
-    for possibility in possibilites:
-      if possibility.is_dir():
-        return possibility
+    def _check_for_standard(self):
+        """ Determine directory that contains the output standard """
+        possibilites = self._example_path.glob('*_Standard')
+        for possibility in possibilites:
+            if possibility.is_dir():
+                return possibility
 
-  def assert_not_stubs(self):
-    assert self._mpi.stubs == 0
+    def assert_not_stubs(self):
+        assert self._mpi.stubs == 0
 
-  def assert_mpi(self):
-    self.assert_not_stubs()
-    assert self._mpi.size >= 1
-    self._mpi.set_added_comm()
+    def assert_mpi(self):
+        self.assert_not_stubs()
+        assert self._mpi.size >= 1
+        self._mpi.set_added_comm()
 
-  def assert_stubs(self):
-    assert self._mpi.stubs == 1
+    def assert_stubs(self):
+        assert self._mpi.stubs == 1
 
-  def run_fitsnap(self):
-    assert importlib.util.find_spec("fitsnap3") is not None
-    original_arguments = sys.argv
-    #new_arguments = ['fitsnap3lib', str(self._input_file), '--overwrite', '-r']
-    new_arguments = ['fitsnap3', str(self._input_file), '--overwrite', '-r']
-    sys.argv = new_arguments
-    run_module(sys.argv[0], run_name='__main__', alter_sys=True)
-    sys.argv = original_arguments
-    
+    def run_fitsnap(self):
+        assert importlib.util.find_spec("fitsnap3") is not None
+        original_arguments = sys.argv
+        #new_arguments = ['fitsnap3lib', str(self._input_file), '--overwrite', '-r']
+        new_arguments = ['fitsnap3', str(self._input_file), '--overwrite', '-r']
+        sys.argv = new_arguments
+        run_module(sys.argv[0], run_name='__main__', alter_sys=True)
+        sys.argv = original_arguments
+      
 
-  def snapcoeff_diff(self):
-    """ Check if current SNAP output differs too much from standardized output"""
-    if self._mpi.rank != 0:
-      return
-    self._outfile += ".snapcoeff"
-    standard_outfile = str(self._standard_path / Path(self._outfile).name)
-    testing_coeffs = self._snap_parser(str(self._example_path / self._outfile))
-    standard_coeffs = self._snap_parser(standard_outfile)
-    assert np.max(testing_coeffs - standard_coeffs) < 1e-6
[…]
