+++ tests/test_examples.py
@@ -1,5 +1,4 @@
 import pytest
-import sys
 from pathlib import Path
 import os
 import importlib.util
