### 2023-03-07 6ec9421
Fix W391 and F811 in fitsnap3lib.tools
