+++ tests/example_checker.py
@@ -2,7 +2,7 @@ import numpy as np
 from pathlib import Path
 import configparser
 import platform
-from subprocess import PIPE, CalledProcessError, run, check_output, list2cmdline
+from subprocess import PIPE, CalledProcessError, run, check_output
 import sys
 from functools import wraps
 import inspect
