+++ fitsnap3lib/units/time.py
@@ -15,4 +15,3 @@ atomic = 1e3
 fs = 1e3
 femto = 1e3
 femtosecond = 1e3
-
