+++ setup.cfg
@@ -1,7 +1,8 @@
 [flake8]
 # Invoke this flake8 setup by calling `flake8` in the terminal.
 max-line-length = 127
-select = E123,E101,W191,E111,E114,E116,E117,F811,F401,W391
+# Seems to be a bug in E111, E117, and E114 indentation multiple of 4, so we will exclude.
+select = E123,E101,W191,E116,F811,F401,W391
 exclude =
     .git,
     __pycache__,
