+++ fitsnap3lib/tools/dataloaders.py
@@ -1,8 +1,6 @@
 import torch.utils.data
-from torch.utils.data import DataLoader
 from sys import float_info
 import numpy as np
-import itertools, operator
 
 
 class InRAMDataset(torch.utils.data.Dataset):
