+++ fitsnap3lib/fitsnap.py
@@ -63,8 +63,11 @@ class FitSnap:
         self.solver = solver(self.config.sections["SOLVER"].solver)
         self.fit = None
         self.multinode = 0
-        self.delete_data = True # make False if don't want to delete data object
-                                # useful for using library to loop over fits
+        """
+        Make False if don't want to delete data object.
+        Useful for using library to loop over fits.
+        """
+        self.delete_data = True
         if self.config.sections["EXTRAS"].only_test:
             self.fit = output.read_fit()
 
