+++ fitsnap3lib/solvers/anl.py
@@ -78,4 +78,3 @@ class ANL(Solver):
         pt = self.pt
         b = pt.shared_arrays['a'].array @ self.fit
         np.savez_compressed('b.npz', b=b)
-
