+++ setup.cfg
@@ -2,13 +2,16 @@
 # Invoke this flake8 setup by calling `flake8` in the terminal.
 max-line-length = 127
 # Seems to be a bug in E111, E117, and E114 indentation multiple of 4, so we will exclude.
-select = E123,E101,W191,E116,F811,F401,W391
+# F401 is required in fitsnap3lib.solvers.solver_factory where we initialize all solvers.
+# This is useful in future for combining solvers.
+select = E123,E101,W191,E116,F811,W391
 exclude =
     .git,
     __pycache__,
     docs,
     build,
     dist,
+    src,
     .eggs,
     *.egg
 max-complexity = 10
