+++ fitsnap3lib/lib/sym_ACE/yamlpace_tools/potential.py
@@ -1,4 +1,3 @@
-import yaml
 import itertools
 from fitsnap3lib.lib.sym_ACE.wigner_couple import *
 from fitsnap3lib.lib.sym_ACE.rpi_lib import *
@@ -7,264 +6,308 @@ import json
 
 
 class AcePot():
-  def __init__(self,
-        elements,
-        reference_ens,
-        ranks,
-        nmax,
-        lmax,
-        nradbase,
-        rcut,
-        lmbda,
-        rcutinner=0.0,
-        drcutinner=0.01,
-        lmin = 1,
-        RPI_heuristic='root_SO3_span',
-        **kwargs):
-    if kwargs != None:
-      self.__dict__.update(kwargs)
-    
-    self.global_ccs = generate_ccs()
-    self.E0 = reference_ens
-    self.ranks =ranks
-    self.elements = elements
-    self.betas = None
-    self.nus = None
-    self.deltaSplineBins=0.001
-    self.global_ndensity=1
-    self.global_FSparams=[1.0, 1.0]
-    self.global_rhocut = 100000
-    self.global_drhocut = 250
-    #assert the same nmax,lmax,nradbase (e.g. same basis) for each bond type
-    self.radbasetype = 'ChebExpCos'
-    self.global_nmax=nmax
-    self.global_lmax=lmax
-    assert len(nmax) == len(lmax),'nmax and lmax arrays must be same size'
-    self.global_nradbase=nradbase
-    if type(rcut) != dict and type(rcut) != list:
-      self.global_rcut = rcut
-      self.global_lmbda = lmbda
-      self.global_rcutinner = rcutinner
-      self.global_drcutinner = drcutinner
-    else:
-      self.rcut = rcut
-      self.lmbda = lmbda
-      self.rcutinner = rcutinner
-      self.drcutinner = drcutinner
-    self.RPI_heuristic = RPI_heuristic
-    self.set_embeddings()
-    self.set_bonds()
-    self.set_bond_base()
-
-    lmax_dict = {rank:lv for rank,lv in zip(self.ranks,self.global_lmax)}
-    nradmax_dict = {rank:nv for rank,nv in zip(self.ranks,self.global_nmax)}
-    mumax_dict={rank:len(self.elements) for rank in self.ranks}
-    if self.RPI_heuristic == 'lexicographic':
-      nulst_1 = [generate_nl(rank,nradmax_dict[rank],lmax_dict[rank],mumax_dict[rank]) for rank in self.ranks]
-    else:
-      nulst_1 = [descriptor_labels_YSG(rank,nradmax_dict[rank],lmax_dict[rank],mumax_dict[rank],lmin) for rank in self.ranks]
-    nus_unsort = [item for sublist in nulst_1 for item in sublist]
-    nus = nus_unsort.copy()
-    mu0s = []
-    mus =[]
-    ns = []
-    ls = []
-    for nu in nus_unsort:
-      mu0ii,muii,nii,lii = get_mu_n_l(nu)
-      mu0s.append(mu0ii)
-      mus.append(tuple(muii))
-      ns.append(tuple(nii))
-      ls.append(tuple(lii))
-    nus.sort(key = lambda x : mus[nus_unsort.index(x)],reverse = False)
-    nus.sort(key = lambda x : ns[nus_unsort.index(x)],reverse = False)
-    nus.sort(key = lambda x : ls[nus_unsort.index(x)],reverse = False)
-    nus.sort(key = lambda x : mu0s[nus_unsort.index(x)],reverse = False)
-    nus.sort(key = lambda x : len(x),reverse = False)
-    nus.sort(key = lambda x : mu0s[nus_unsort.index(x)],reverse = False)
[…]
