+++ setup.cfg
@@ -12,6 +12,7 @@ exclude =
     build,
     dist,
     src,
+    examples,
     .eggs,
     *.egg
 max-complexity = 10

[…]
