+++ setup.cfg
@@ -1,10 +1,9 @@
 [flake8]
 # Invoke this flake8 setup by calling `flake8` in the terminal.
 max-line-length = 127
-# Seems to be a bug in E111, E117, and E114 indentation multiple of 4, so we will exclude.
 # F401 is required in fitsnap3lib.solvers.solver_factory where we initialize all solvers.
 # This is useful in future for combining solvers.
-select = E123,E101,W191,E116,F811,W391
+select = E123,E101,W191,E116,F811,W391,E111
 exclude =
     .git,
     __pycache__,
