### 2023-03-08 #5118 @rohskopf
Hi @tpurcell90 and @jmschrei, quick update.

We addressed the issues @tpurcell90 raised, outlined below. The issues also show which PRs and commits were involved.

https://github.com/FitSNAP/FitSNAP/issues/180 : We added a `pyproject.toml` file and uploaded our package to PyPI, so that users can install with pip.

https://github.com/FitSNAP/FitSNAP/issues/181 : We have an internal function that converts settings to the correct data type internally, regardless of whether 1/0 or True/False is used for settings.

https://github.com/FitSNAP/FitSNAP/issues/182 : I added `flake8` linting in `setup.cfg` for formatting specs that we care about, and added a note for contributors about using `flake8` to verify proper format.

https://github.com/FitSNAP/FitSNAP/issues/183 : Our CI tests use OpenMPI so we never caught this. We use the tests to verify that the code gives consistent results and that our interatomic potentials are correctly coded.

https://github.com/FitSNAP/FitSNAP/issues/184 : Colab tutorial is fixed, just required minor changes to use new NN file formats.

https://github.com/FitSNAP/FitSNAP/issues/185 : Since LAMMPS warnings are useful for developers at this time, and sometimes rely on external packages, it is best to keep these warnings. I therefore mentioned this in the installation docs so that users are not alarmed.

https://github.com/FitSNAP/FitSNAP/issues/186 : We have an example using our library combined with ASE, which also has scrapers. We prefer this instead of making ASE another dependency.

On the remaining checklist items for functionality:

I will note that the installation procedure is used in the Colab tutorial, but we also have `pip` as an option now. As for functional claims, the purpose of our software is to learn a mapping of atomic positions to energies and forces, then deploy these models for simulation in LAMMPS. I think our tutorial illustrates this nicely. Our CI tests for NN potentials also confirm that forces are correctly coded by comparing analytical forces to finite difference forces.

Thank you @tpurcell90 for the diligent and descriptive issues that were raised!
