### 2023-03-09 #5118 @tpurcell90
HI Drew, I'll take a look at the collab notebook once I am back from APS (next week). Assuming that everything works with OpenMPI then most of my concerns will likely be addressed and I can complete the check list.

The intel MPI issue and collab notebook problems were the two main reasons I could not check them off.
