### 2023-03-10 #5118 @rohskopf
Thanks @bahung and @tpurcell90 .
Tutorial was fixed iand merged in https://github.com/FitSNAP/FitSNAP/pull/192
