### 2023-03-13 FitSNAP/FitSNAP#193 @bahung
JOSS Review example for uncertainty qualification (UQ)
This is part of the review process for FitSNAP [Issue 5118.](https://github.com/openjournals/joss-reviews/issues/5118)

Please incorporate an example for uncertainty qualification (UQ) to the colab notebook demonstrating the uncertainty quantification (UQ) capabilities of FitSNAP.
