### 2023-03-13 #5118 @bahung
My comments are as follows after the colab installation and running the author's provided colab notebook:

1. Installation: It has some 'declared with attribute warn_unused_result' warnings which are mainly from LAMMPS, I think we better have some notes in installation and/or documentation as @tpurcell90 has already mentioned
2. Functionality: Most of the claims of FitSNAP are satisfied. It could be great if the uncertainty quantification (UQ), while described in the documentation, is also illustrated in the colab notebook. Thus I cannot verify the full functionality at the moment. I created an issue here: https://github.com/FitSNAP/FitSNAP/issues/193
3. Documentation: Section 4 Linear models with only one subsection seems a bit unnormal, please consider adding/separating how we fit linear models and so forth.
4. Paper: I think the reference for [janssen2019pyiron] is missing, please include it in the paper.
5. Tutorial: It has a detailed tutorial in Google Colab which is fine, it could be great if the colab notebook is more well-organised to present all FitSNAP functionalities. Eg. consistency of numberings, functions, or real-world examples.

Thanks,
Ba Hung
