### 2023-03-14 #5118 @rohskopf
Thank you @bahung for these excellent points.

I will address them one by one as separate comments below, and make necessary changes in https://github.com/FitSNAP/FitSNAP/pull/194
