### 2023-03-14 #5118 @rohskopf
> 1. Installation: It has some 'declared with attribute warn_unused_result' warnings which are mainly from LAMMPS, I think we better have some notes in installation and/or documentation as @tpurcell90 has already mentioned

I added a small comment to the docs to tell users not to be warned: "Do not be alarmed by runtime library warnings after cmake, or -Weffc++ and -Wunused-result warnings during make." Fixed and explained in https://github.com/FitSNAP/FitSNAP/issues/185
