### 2023-03-14 #5118 @rohskopf
>2. Functionality: Most of the claims of FitSNAP are satisfied. It could be great if the uncertainty quantification (UQ), while described in the documentation, is also illustrated in the colab notebook. Thus I cannot verify the full functionality at the moment.

Thanks for pointing out the need for a UQ illustration. I added a UQ example to the colab notebook, in a section like this:

This outputs a `covariance.npy` file that has covariance values for each fitting coefficient. I show in the tutorial that we can make parity plots with these uncertainties, e.g. here's a plot showing comparisons in energy between model and target:

I also added a more appropriate UQ section in the docs: https://fitsnap.github.io/Linear.html#uncertainty-quantification-uq

Here we explain the UQ method that we have implemented, an analytical Bayesian method that produces the output covariance values. We reserve a separate section for more experimental/exploratory methods here: https://fitsnap.github.io/Linear.html#in-development-uq-solvers

In the paper, we mention that our UQ solvers output covariances for each fitting coefficient, and that is what we added to the tutorial with the former method.
