### 2023-03-14 #5118 @rohskopf
> 3. Documentation: Section 4 Linear models with only one subsection seems a bit unnormal, please consider adding/separating how we fit linear models and so forth.

Thank you for pointing this out. I moved the UQ method to the linear section since our UQ only applies to linear models for now. We also show the exact fitting matrix we use, e.g. shown here:
![image](https://user-images.githubusercontent.com/11083811/224908517-c1dd3e4c-a716-42bc-baca-232c4a1609bf.png)

Then we also state:

> FitSNAP solves this matrix problem using SVD and supports other solvers as well. For more details on settings used for descriptors and solvers, please see the docs on [FitSNAP Input Scripts](file:///Users/adrohsk/FitSNAP/docs/build/html/Run.html#input-scripts).
