### 2023-03-14 #5118 @rohskopf
> 4. Paper: I think the reference for [janssen2019pyiron] is missing, please include it in the paper.

Thanks for catching. Fixed.
