### 2023-03-14 #5118 @rohskopf
> 5. Tutorial: It has a detailed tutorial in Google Colab which is fine, it could be great if the colab notebook is more well-organised to present all FitSNAP functionalities. Eg. consistency of numberings, functions, or real-world examples.

Thanks for bringing this up. Now that UQ is added to the tutorial this completes all of our current functionalities. The rest of the examples are the same, but with different training data.

The purpose of our software is to learn a mapping between atomic positions and energies/forces extracted from quantum mechanics. Then deploy these models for running molecular dynamics simulations. The tutorial illustrates this purpose for some example materials involved in nuclear energy applications (these were also cited in the paper).

There are many real-world examples of applying molecular dynamics simulations to a variety of problems, such as calculating/predicting material properties which is very useful in materials engineering. These are best left to the paper with citations pointing towards the various publications that have used FitSNAP and/or molecular dynamics simulations.
