### 2023-03-14 #5118 @rohskopf
> 5. Tutorial: It has a detailed tutorial in Google Colab which is fine, it could be great if the colab notebook is more well-organised to present all FitSNAP functionalities. Eg. consistency of numberings, functions, or real-world examples.

More comments regarding consistency of numbers/results and functions:

To further satsify this, we added a Tests section to the docs which explains our CI tests: https://fitsnap.github.io/Tests.html

As seen there, our CI tests verify:

1. The consistency of fitting coefficients from our linear solvers, using results from older versions of our software, and
2. The correctness of neural network derivatives w.r.t. atomic positions. These are the atomic forces that we fit to. We verify these with finite difference calculations.

To further illustrate the correctness of neural network derivatives w.r.t. atomic positions, I added a section to the tutorial here:

which shows a distribution of errors between analytical/autodiff forces and finite difference forces:
![image](https://user-images.githubusercontent.com/11083811/224911684-8831b85a-b4f2-49ba-850d-59e378227cb4.png)
