### 2023-03-18 #5118 @bahung
Thank you @rohskopf for the promtly responses.
The functionalities of FitSNAP are now confirmed with elaborated examples. I completed my checklist.
