### 2023-03-21 FitSNAP/FitSNAP#193 @rohskopf
Fixed in https://github.com/FitSNAP/FitSNAP/pull/194
