### 2023-03-23 ccb36a4
Write pickled configuration list
