### 2023-03-23 #5118 @rohskopf
@jmschrei we don't have a DOI for the code+paper, is that required?

Version number is 3.0.1

Fixing DOI now.
