### 2023-03-23 #5118 @jmschrei
Yes, a DOI is required. You can just download a zip file of the release (with the paper in it) and put it on Zenodo, though.
