### 2023-03-27 8204fb8
Add keyword option for dumping pickled list of configs
