### 2023-03-27 #5118 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
