+++ examples/UO2_PACE/uo2_nocharge.in
@@ -1,66 +1,0 @@
-[ACE]
-numTypes = 2
-rcutfac = 5.0 5.5 5.5 6.0
-lambda = 3.3 3.3 3.3 3.3
-rcinner = 0.1 0.1 0.1 0.1
-drcinner = 0.01 0.01 0.01 0.01
-ranks = 1 2 3 4 
-lmax =  1 2 2 1
-nmax = 16 2 2 1
-mumax = 2
-nmaxbase = 16
-type = O U
-bzeroflag = 0
-
-[CALCULATOR]
-calculator = LAMMPSPACE
-energy = 1
-force = 1
-stress = 0
-
-[ESHIFT]
-O = 0.0
-U = 0.0
-
-[SOLVER]
-solver = ARD
-compute_testerrs = 1
-detailed_errors = 1
-[ARD]
-alphabig=1e-6
-alphasmall=1e-6
-lambdabig=1e-6
-lambdasmall=1e-6
-threshold_lambda=10000
-
-[SCRAPER]
-scraper = JSON
-
-[PATH]
-dataPath = JSON
-
-[OUTFILE]
-output_style = PACE
-metrics = UO2_metrics.md
-potential = UO2_pot
-
-[REFERENCE]
-units = metal
-atom_style = atomic
-pair_style = hybrid/overlay zero 6.0
-pair_coeff1 = * * zero
-
-[GROUPS]
-group_sections = name training_size testing_size eweight fweight vweight
-group_types = str float float float float float
-smartweights = 0
-random_sampling = 0
-latte_md	=	0.2	0	1	1.0	1.00E-12
-
-[EXTRAS]
-dump_descriptors = 0
-dump_truth = 0
-dump_weights = 0
-
-[MEMORY]
-override = 0

[…]
