+++ examples/Ta_PACE/Ta.in
@@ -1,17 +1,15 @@
 [ACE]
 numTypes = 1
-rcutfac = 4.25
-lambda = 1.275
-rcinner = 1.2
-drcinner = 0.01
-ranks = 1 2 3 4
-lmax =  0 5 2 1
-nmax =  22 5 3 1
-mumax = 1
-lmin = 0 0 1 1
+ranks = 1 2 3 4 5 6
+lmax =  1 2 2 2 1 1
+nmax = 22 2 2 2 1 1
 nmaxbase = 22
+rcutfac = 4.604694451
+lambda = 3.059235105
 type = Ta
+lmin = 1
 bzeroflag = 0
+RPI_heuristic = root_SO3_span
 
 [CALCULATOR]
 calculator = LAMMPSPACE
@@ -23,19 +21,15 @@ stress = 0
 Ta = 0.0
 
 [SOLVER]
-solver = RIDGE
+solver = SVD
 compute_testerrs = 1
 detailed_errors = 1
-
-[RIDGE]
-local_solver = 1
-alpha = 1.e-4
 
 [SCRAPER]
 scraper = JSON
 
 [PATH]
-dataPath = /home/jmgoff/work/new_Ta_ex/single_elm_grid/JSON
+dataPath = ../Ta_Linear_JCP2014/JSON
 
 [OUTFILE]
 output_style = PACE
@@ -45,31 +39,33 @@ potential = Ta_pot
 [REFERENCE]
 units = metal
 atom_style = atomic
-pair_style = hybrid/overlay zero 4.25 
+pair_style = hybrid/overlay zero 5.0 zbl 4.0 4.8
 pair_coeff1 = * * zero
+pair_coeff2 = * * zbl 73 73
 
 [GROUPS]
+# name size eweight fweight vweight
 group_sections = name training_size testing_size eweight fweight vweight
 group_types = str float float float float float
-smartweights = 1
+smartweights = 0
 random_sampling = 0
-Displaced_A15 =  1.0    0.0           1. 20. 1.E-8
-Displaced_BCC =  1.0    0.0           1. 20. 1.E-8
-Displaced_FCC =  1.0    0.0           1. 20. 1.E-8
-Elastic_BCC   =  1.0    0.0           1. 10. 1.E-8
-Elastic_FCC   =  1.0    0.0           1. 10. 1.E-8
-GSF_110       =  1.0    0.0           1. 10. 1.E-8
-GSF_112       =  1.0    0.0           1. 10. 1.E-8
-Liquid        =  1.0    0.0           1. 10. 1.E-8
-Surface       =  1.0    0.0           1. 10. 1.E-8
-Volume_A15    =  1.0    0.0           1.5 15. 1.E-8
-Volume_BCC    =  1.0    0.0           1.5 15. 1.E-8
-Volume_FCC    =  1.0    0.0           1.5 15. 1.E-8
+Displaced_A15 =  1.0    0.0       28.78562746  0.9985573607     1.00E-08
+Displaced_BCC =  1.0    0.0       28.78562746  0.9985573607     1.00E-08
+Displaced_FCC =  1.0    0.0       28.78562746  0.9985573607     1.00E-08
+Elastic_BCC   =  1.0    0.0       115.5696535    0.39682072       1.00E-08
+Elastic_FCC   =  1.0    0.0       115.5696535    0.39682072       1.00E-08
+GSF_110       =  1.0    0.0       30.93231609       2.049246789           1.00E-08
+GSF_112       =  1.0    0.0       30.93231609        2.049246789           1.00E-08
+Liquid        =  1.0    0.0       274.4904041       1.439344999        1.00E-08
+Surface       =  1.0    0.0       23.07342458    1.562790276       1.00E-08
[…]
