### 2023-04-03 cbdb9b3
updated ACE examples with regularization
