+++ examples/Ta_PACE_RIDGE/30Mar23_RIDGE/Ta_metrics.md
@@ -1,0 +1,54 @@
+|                                                       |   ncount |         mae |        rmse |          rsq |
+|:------------------------------------------------------|---------:|------------:|------------:|-------------:|
+| ('*ALL', 'Unweighted', 'Training', 'Energy')          |      363 | 0.0549201   | 0.0774879   |     0.99953  |
+| ('*ALL', 'Unweighted', 'Training', 'Force')           |    12672 | 0.0664154   | 0.122804    |     0.969273 |
+| ('*ALL', 'weighted', 'Training', 'Energy')            |      363 | 0.00210791  | 0.0038682   |     0.999943 |
+| ('*ALL', 'weighted', 'Training', 'Force')             |    12672 | 0.000748996 | 0.00142301  |     0.973678 |
+| ('Displaced_A15', 'Unweighted', 'Training', 'Energy') |        9 | 0.021013    | 0.021021    |   -58.5354   |
+| ('Displaced_A15', 'Unweighted', 'Training', 'Force')  |     1728 | 0.0772245   | 0.0984937   |     0.976044 |
+| ('Displaced_A15', 'weighted', 'Training', 'Energy')   |        9 | 0.00233478  | 0.00233566  |   -58.5354   |
+| ('Displaced_A15', 'weighted', 'Training', 'Force')    |     1728 | 0.000893802 | 0.00113997  |     0.976044 |
+| ('Displaced_BCC', 'Unweighted', 'Training', 'Energy') |        9 | 0.0318285   | 0.0318592   |     0.575502 |
+| ('Displaced_BCC', 'Unweighted', 'Training', 'Force')  |     1458 | 0.111004    | 0.137839    |     0.989391 |
+| ('Displaced_BCC', 'weighted', 'Training', 'Energy')   |        9 | 0.0035365   | 0.00353991  |     0.575502 |
+| ('Displaced_BCC', 'weighted', 'Training', 'Force')    |     1458 | 0.00152269  | 0.00189079  |     0.989391 |
+| ('Displaced_FCC', 'Unweighted', 'Training', 'Energy') |        9 | 0.0701627   | 0.0701654   |  -305.588    |
+| ('Displaced_FCC', 'Unweighted', 'Training', 'Force')  |     1296 | 0.0890234   | 0.112026    |     0.950547 |
+| ('Displaced_FCC', 'weighted', 'Training', 'Energy')   |        9 | 0.00779586  | 0.00779615  |  -305.588    |
+| ('Displaced_FCC', 'weighted', 'Training', 'Force')    |     1296 | 0.00137382  | 0.0017288   |     0.950547 |
+| ('Elastic_BCC', 'Unweighted', 'Training', 'Energy')   |      100 | 0.0192154   | 0.0192163   | -1795.74     |
+| ('Elastic_BCC', 'Unweighted', 'Training', 'Force')    |      600 | 1.30835e-05 | 1.79681e-05 |     0.99701  |
+| ('Elastic_BCC', 'weighted', 'Training', 'Energy')     |      100 | 0.000192154 | 0.000192163 | -1795.74     |
+| ('Elastic_BCC', 'weighted', 'Training', 'Force')      |      600 | 2.18058e-07 | 2.99469e-07 |     0.99701  |
+| ('Elastic_FCC', 'Unweighted', 'Training', 'Energy')   |      100 | 0.0700236   | 0.0700278   | -1202.36     |
+| ('Elastic_FCC', 'Unweighted', 'Training', 'Force')    |     1200 | 9.77305e-05 | 0.000143144 |     0.874894 |
[…]
