+++ examples/Ta_PACE/auto_fit_this.py
@@ -23,7 +23,7 @@ def change_descriptor_hyperparams(config):
 	config.sections['ACE'].rcinner = ['0.0']
 	config.sections['ACE'].drcinner = ['0.01']
 
-	config.sections['ACE'].lmin = 1
+	config.sections['ACE'].lmin = [1]
 	config.sections['ACE'].ranks = [1, 2, 3, 4, 5, 6]
 	config.sections['ACE'].lmax = [1, 2, 2, 2, 1, 1]
 	config.sections['ACE'].nmax =[22, 2, 2, 2, 1, 1]
