### 2023-04-04 d38f42c
Fix class name exception in custom calculator
