### 2023-04-04 59e55fd
fix training group name extraction from input file
This fixes a small bug with pathed group names in FitSNAP input files
Previous behavior took the second-to-last part of the split filename, leading to potential duplicate dictionary keys despite unique path names.
Current behavior takes everything between the dataPath variable and the JSON file name, leaving just the pathed group name and ensuring unique dictionary keys.
