### 2023-04-05 #5118 @Kevin-Mattheus-Moerman
@rohskopf I am the AEiC on this track and here to help process the final steps. For the moment can you check the following:

- [ ] Only the first author shows with an ORCID id. Can you check if you can add more/all? Note we also recommend that you add the ORCID information on the ZENODO archive.
- [ ] Please update the license listed on the ZENODO archive to match you software's open source license.
- [ ] Your repository does not currently seem to list a release with the version tag `3.0.1`. Can you please create one.
