### 2023-04-05 abe81ba
clean up scraper comment lines
