+++ fitsnap3lib/scrapers/vasp_scraper.py
@@ -40,18 +40,18 @@ class Vasp(Scraper):
 
 
     def scrape_groups(self):
-        ### Locate all OUTCARs in datapath
+        # Locate all OUTCARs in datapath
         glob_asterisks = '/**/*'
         outcars_base = os.path.join(self.vasp_path, *glob_asterisks.split('/'))
 
-        ## TODO make this search user-specify-able (e.g., OUTCARs have labels/prefixes etc.)
+        # TODO make this search user-specify-able (e.g., OUTCARs have labels/prefixes etc.)
         all_outcars = [f for f in glob(outcars_base,recursive=True) if f.endswith('OUTCAR')]
 
-        ## Grab test|train split
+        # Grab test|train split
         self.group_dict = {k: config.sections['GROUPS'].group_types[i] for i, k in enumerate(config.sections['GROUPS'].group_sections)}
 
         for group in self.group_table:
-            ## First, check that all group folders exist
+            # First, check that all group folders exist
             group_vasp_path = f'{self.vasp_path}/{group}'
             if not os.path.exists(group_vasp_path):
                 raise Exception('!!ERROR: group folder not detected!!' 
@@ -77,8 +77,8 @@ class Vasp(Scraper):
             if training_size is None:
                 raise ValueError('Please set training size for {}'.format(group))
             
-            ## Grab OUTCARS for this training group
-            ## Test filepath to be sure that unique group name is being matched
+            # Grab OUTCARS for this training group
+            # Test filepath to be sure that unique group name is being matched
             group_outcars = [f for f in all_outcars if group_vasp_path + '/' in f]
             if len(group_outcars) == 0:
                 raise Exception('!!ERROR: no OUTCARs found in group!!' 
@@ -92,56 +92,56 @@ class Vasp(Scraper):
             self.configs[group] = []  
 
             for outcar in self.files[file_base]:
-                ## Open file
+                # Open file
                 with open(outcar, 'r') as fp:
                     lines = fp.readlines()
                 nlines = len(lines)
 
-                ## Use ion loop text to partition ionic steps
+                # Use ion loop text to partition ionic steps
                 ion_loop_text = 'aborting loop'
                 start_idx_loops = [i for i, line in enumerate(lines) if ion_loop_text in line]
                 converged_list = [False if 'unconverged' in lines[i] else True for i in start_idx_loops]
                 end_idx_loops = [i for i in start_idx_loops[1:]] + [nlines]
 
-                ## Grab potcar and element info
+                # Grab potcar and element info
                 header_lines = lines[:start_idx_loops[0]]
                 potcar_list, potcar_elements, ions_per_type, is_duplicated = self.parse_outcar_header(header_lines)
 
-                ## Each config in a single OUTCAR is assigned the same
-                ## parent data (i.e. filename, potcar and ion data)
-                ## but separated for each iteration (idx loops on 'lines')
-                ## Tuple data: outcar file name str, config number int, starting line number (for debug)  int, 
-                ## potcar list, potcar elements list, number ions per element list, configuration lines list 
+                # Each config in a single OUTCAR is assigned the same
+                # parent data (i.e. filename, potcar and ion data)
+                # but separated for each iteration (idx loops on 'lines')
+                # Tuple data: outcar file name str, config number int, starting line number (for debug)  int, 
+                # potcar list, potcar elements list, number ions per element list, configuration lines list 
                 unique_configs = [(outcar, i, start_idx_loops[i], potcar_list, potcar_elements, ions_per_type, converged_list[i],
[…]
